Support for reading SBML3 files with FBC2 extension.
Adding an 'SBMLResourceIdentifiers' class for importing references on Genes, Proteins, Metabolites, and Reactions.  SBML importer can now read geneProduct FBC objects and turn them into Protein objects.  Importer also reads “is” and “isEncodedBy” MIRIAM biological qualifiers and parses them into SBMLResourceIdentifiers.

Correctly reading the fbc:upperFluxBound and fbc:lowerFluxBound values.  Clarifying some exceptions in rangedNumber, Proteins, and Genes.  Additional checking for SBML import to work nicely with versions of the library lacking FBC.  Extracting chemical formula from FBC SBML attribute for metabolites.  Adding code to track 'non-canonical' proteins - i.e. proteins that represent unknown connectors between genes and reactions.  Untangling a series of circular import dependencies that prevented modular testing of parts of the JQMM including sbmlio.
