+++ code/core/MetaboliteFormulaDictionary.py
@@ -1,0 +1,1835 @@
+# Part of the JBEI Quantitative Metabolic Modeling Library (JQMM)
+# Copyright (c) 2016, The Regents of the University of California.
+# For licensing details see "license.txt" and "legal.txt".
+
+
+formulaDict = {
+    '10fthf': 'C20H21N7O7',
+    '12dgr120': 'C27H52O5',
+    '12dgr120_p': 'C27H52O5',
+    '12dgr140': 'C31H60O5',
+    '12dgr140_p': 'C31H60O5',
+    '12dgr141': 'C31H56O5',
+    '12dgr141_p': 'C31H56O5',
+    '12dgr160': 'C35H68O5',
+    '12dgr160_p': 'C35H68O5',
+    '12dgr161': 'C35H64O5',
+    '12dgr161_p': 'C35H64O5',
+    '12dgr180': 'C39H76O5',
+    '12dgr180_p': 'C39H76O5',
+    '12dgr181': 'C39H72O5',
+    '12dgr181_p': 'C39H72O5',
+    '12ppd_R': 'C3H8O2',
+    '12ppd_R_e': 'C3H8O2',
+    '12ppd_R_p': 'C3H8O2',
+    '12ppd_S': 'C3H8O2',
+    '12ppd_S_e': 'C3H8O2',
+    '12ppd_S_p': 'C3H8O2',
+    '13dpg': 'C3H4O10P2',
+    '14dhncoa': 'C32H38N7O19P3S',
+    '14glucan': 'C36H62O31',
+    '14glucan_e': 'C36H62O31',
+    '14glucan_p': 'C36H62O31',
+    '15dap': 'C5H16N2',
+    '15dap_e': 'C5H16N2',
+    '15dap_p': 'C5H16N2',
+    '1agpe120_p': 'C17H36NO7P1',
+    '1agpe140_p': 'C19H40NO7P1',
+    '1agpe141_p': 'C19H38NO7P1',
+    '1agpe160_p': 'C21H44NO7P1',
+    '1agpe161_p': 'C21H42NO7P1',
+    '1agpe180_p': 'C23H48NO7P1',
+    '1agpe181_p': 'C23H46NO7P1',
+    '1agpg120_p': 'C18H36O9P1',
+    '1agpg140_p': 'C20H40O9P1',
+    '1agpg141_p': 'C20H38O9P1',
+    '1agpg160_p': 'C22H44O9P1',
+    '1agpg161_p': 'C22H42O9P1',
+    '1agpg180_p': 'C24H48O9P1',
+    '1agpg181_p': 'C24H46O9P1',
+    '1ddecg3p': 'C15H29O7P1',
+    '1ddecg3p_p': 'C15H29O7P1',
+    '1hdec9eg3p': 'C19H35O7P1',
+    '1hdec9eg3p_p': 'C19H35O7P1',
+    '1hdecg3p': 'C19H37O7P1',
+    '1hdecg3p_p': 'C19H37O7P1',
+    '1odec11eg3p': 'C21H39O7P1',
+    '1odec11eg3p_p': 'C21H39O7P1',
+    '1odecg3p': 'C21H41O7P1',
+    '1odecg3p_p': 'C21H41O7P1',
+    '1pyr5c': 'C5H6NO2',
+    '1tdec7eg3p': 'C17H31O7P1',
+    '1tdec7eg3p_p': 'C17H31O7P1',
+    '1tdecg3p': 'C17H33O7P1',
+    '1tdecg3p_p': 'C17H33O7P1',
+    '23camp_e': 'C10H11N5O6P',
+    '23camp_p': 'C10H11N5O6P',
+    '23ccmp_e': 'C9H11N3O7P',
+    '23ccmp_p': 'C9H11N3O7P',
+    '23cgmp_e': 'C10H11N5O7P',
+    '23cgmp_p': 'C10H11N5O7P',
+    '23cump_e': 'C9H10N2O8P',
+    '23cump_p': 'C9H10N2O8P',
+    '23dappa': 'C3H9N2O2',
+    '23dappa_e': 'C3H9N2O2',
+    '23dappa_p': 'C3H9N2O2',
+    '23ddhb': 'C7H7O4',
+    '23dhacoa': 'C27H37N7O19P3S',
+    '23dhb': 'C7H5O4',
+    '23dhba': 'C17H17N5O10P',
+    '23dhbzs': 'C10H10NO6',
+    '23dhdp': 'C7H5NO4',
+    '23dhmb': 'C5H9O4',
+    '23dhmp': 'C6H11O4',
+    '23doguln': 'C6H7O7',
+    '25aics': 'C13H15N4O12P',
+    '25dkglcn': 'C6H7O7',
+    '25drapp': 'C9H14N5O8P',
+    '26dap_LL': 'C7H14N2O4',
+    '26dap_M': 'C7H14N2O4',
+    '26dap_M_e': 'C7H14N2O4',
+    '26dap_M_p': 'C7H14N2O4',
+    '2agpe120': 'C17H36NO7P1',
+    '2agpe120_p': 'C17H36NO7P1',
+    '2agpe140': 'C19H40NO7P1',
+    '2agpe140_p': 'C19H40NO7P1',
+    '2agpe141': 'C19H38NO7P1',
+    '2agpe141_p': 'C19H38NO7P1',
+    '2agpe160': 'C21H44NO7P1',
+    '2agpe160_p': 'C21H44NO7P1',
+    '2agpe161': 'C21H42NO7P1',
+    '2agpe161_p': 'C21H42NO7P1',
+    '2agpe180': 'C23H48NO7P1',
+    '2agpe180_p': 'C23H48NO7P1',
+    '2agpe181': 'C23H46NO7P1',
+    '2agpe181_p': 'C23H46NO7P1',
+    '2agpg120': 'C18H36O9P1',
+    '2agpg120_p': 'C18H36O9P1',
+    '2agpg140': 'C20H40O9P1',
+    '2agpg140_p': 'C20H40O9P1',
+    '2agpg141': 'C20H38O9P1',
+    '2agpg141_p': 'C20H38O9P1',
+    '2agpg160': 'C22H44O9P1',
+    '2agpg160_p': 'C22H44O9P1',
+    '2agpg161': 'C22H42O9P1',
+    '2agpg161_p': 'C22H42O9P1',
+    '2agpg180': 'C24H48O9P1',
+    '2agpg180_p': 'C24H48O9P1',
+    '2agpg181': 'C24H46O9P1',
+    '2agpg181_p': 'C24H46O9P1',
+    '2ahbut': 'C6H9O4',
+    '2amsa': 'C3H5NO3',
+    '2aobut': 'C4H7NO3',
+    '2cpr5p': 'C12H13NO9P',
+    '2dda7p': 'C7H10O10P',
+    '2ddecg3p': 'C15H30O7P1',
+    '2ddecg3p_p': 'C15H30O7P1',
+    '2ddg6p': 'C6H8O9P',
+    '2ddglcn': 'C6H9O6',
+    '2ddglcn_e': 'C6H9O6',
+    '2ddglcn_p': 'C6H9O6',
+    '2dh3dgal': 'C6H9O6',
+    '2dh3dgal6p': 'C6H8O9P',
+    '2dhguln': 'C6H9O7',
+    '2dhp': 'C6H9O4',
+    '2dmmq8': 'C50H70O2',
+    '2dmmql8': 'C50H72O2',
+    '2dr1p': 'C5H9O7P',
+    '2dr5p': 'C5H9O7P',
+    '2fe1s': 'SFe2',
+    '2fe2s': 'S2Fe2',
+    '2h3oppan': 'C3H3O4',
+    '2hdec9eg3p': 'C19H36O7P1',
+    '2hdec9eg3p_p': 'C19H36O7P1',
+    '2hdecg3p': 'C19H38O7P1',
+    '2hdecg3p_p': 'C19H38O7P1',
+    '2hone': 'C7H14O',
+    '2hone_e': 'C7H14O',
+    '2hpdeone': 'C17H32O',
+    '2hpdeone_e': 'C17H32O',
+    '2hpdone': 'C17H34O',
+    '2hpdone_e': 'C17H34O',
+    '2ippm': 'C7H8O4',
+    '2mahmp': 'C6H8N3O7P2',
+    '2mcacn': 'C7H5O6',
+    '2mcit': 'C7H7O7',
+    '2me4p': 'C5H11O7P',
+    '2mecdp': 'C5H10O9P2',
+    '2none': 'C9H18O',
+    '2none_e': 'C9H18O',
+    '2obut': 'C4H5O3',
+    '2odec11eg3p': 'C21H40O7P1',
+    '2odec11eg3p_p': 'C21H40O7P1',
+    '2odecg3p': 'C21H42O7P1',
+    '2odecg3p_p': 'C21H42O7P1',
+    '2ohph': 'C46H70O2',
+    '2ombzl': 'C47H72O3',
+    '2omhmbl': 'C48H74O4',
+    '2ommbl': 'C48H74O3',
+    '2omph': 'C47H72O2',
+    '2oph': 'C46H70O',
+    '2oxpaccoa': 'C29H38N7O18P3S',
+    '2p4c2me': 'C14H22N3O17P3',
+    '2pdeone': 'C15H28O',
+    '2pdeone_e': 'C15H28O',
+    '2pdone': 'C15H30O',
+    '2pdone_e': 'C15H30O',
+    '2pg': 'C3H4O7P',
+    '2pglyc': 'C2H2O6P',
+    '2pone': 'C5H10O',
+    '2pone_e': 'C5H10O',
+    '2sephchc': 'C14H13O9',
+    '2shchc': 'C11H10O6',
+    '2tdec7eg3p': 'C17H32O7P1',
+    '2tdec7eg3p_p': 'C17H32O7P1',
+    '2tdecg3p': 'C17H34O7P1',
+    '2tdecg3p_p': 'C17H34O7P1',
+    '2tdeone': 'C13H24O',
+    '2tdeone_e': 'C13H24O',
+    '2tdone': 'C13H26O',
+    '2tdone_e': 'C13H26O',
+    '2tpr3dpcoa': 'C26H40N7O26P5S',
+    '2uone': 'C11H22O',
+    '2uone_e': 'C11H22O',
+    '34dhpac_e': 'C8H8O3',
+    '34dhpac_p': 'C8H8O3',
+    '34hpp': 'C9H7O4',
+    '35cgmp': 'C10H11N5O7P',
+    '3amac': 'C3H4NO2',
+    '3amp_e': 'C10H12N5O7P',
+    '3amp_p': 'C10H12N5O7P',
+    '3c2hmp': 'C7H10O5',
+    '3c3hmp': 'C7H10O5',
+    '3c4mop': 'C7H8O5',
+    '3cmp_e': 'C9H12N3O8P',
+    '3cmp_p': 'C9H12N3O8P',
+    '3dhguln': 'C6H9O7',
+    '3dhgulnp': 'C6H8O10P',
+    '3dhq': 'C7H9O6',
+    '3dhsk': 'C7H7O5',
+    '3fe4s': 'S4Fe3',
+    '3gmp_e': 'C10H12N5O8P',
+    '3gmp_p': 'C10H12N5O8P',
+    '3haACP': 'C15H27N2O9PRS',
+    '3hadpcoa': 'C27H39N7O20P3S',
+    '3hbcoa': 'C25H38N7O18P3S',
+    '3hcddec5eACP': 'C23H41N2O9PRS',
+    '3hcinnm': 'C9H7O3',
+    '3hcinnm_e': 'C9H7O3',
+    '3hcinnm_p': 'C9H7O3',
+    '3hcmrs7eACP': 'C25H45N2O9PRS',
+    '3hcpalm9eACP': 'C27H49N2O9PRS',
+    '3hcvac11eACP': 'C29H53N2O9PRS',
+    '3hdcoa': 'C31H50N7O18P3S',
+    '3hddcoa': 'C33H54N7O18P3S',
+    '3hddecACP': 'C23H43N2O9PRS',
+    '3hdecACP': 'C21H39N2O9PRS',
+    '3hhcoa': 'C27H42N7O18P3S',
+    '3hhdcoa': 'C37H62N7O18P3S',
+    '3hhexACP': 'C17H31N2O9PRS',
+    '3hmrsACP': 'C25H47N2O9PRS',
+    '3hocoa': 'C29H46N7O18P3S',
+    '3hoctaACP': 'C29H55N2O9PRS',
+    '3hoctACP': 'C19H35N2O9PRS',
+    '3hodcoa': 'C39H66N7O18P3S',
+    '3hpalmACP': 'C27H51N2O9PRS',
+    '3hpp': 'C3H5O3',
+    '3hpp_e': 'C3H5O3',
+    '3hpp_p': 'C3H5O3',
+    '3hpppn': 'C9H9O3',
+    '3hpppn_e': 'C9H9O3',
+    '3hpppn_p': 'C9H9O3',
+    '3htdcoa': 'C35H58N7O18P3S',
+    '3ig3p': 'C11H12NO6P',
+    '3mob': 'C5H7O3',
+    '3mop': 'C6H9O3',
+    '3ocddec5eACP': 'C23H39N2O9PRS',
+    '3ocmrs7eACP': 'C25H43N2O9PRS',
+    '3ocpalm9eACP': 'C27H47N2O9PRS',
+    '3ocvac11eACP': 'C29H51N2O9PRS',
+    '3odcoa': 'C31H48N7O18P3S',
+    '3oddcoa': 'C33H52N7O18P3S',
+    '3oddecACP': 'C23H41N2O9PRS',
+    '3odecACP': 'C21H37N2O9PRS',
+    '3ohcoa': 'C27H40N7O18P3S',
+    '3ohdcoa': 'C37H60N7O18P3S',
+    '3ohexACP': 'C17H29N2O9PRS',
+    '3omrsACP': 'C25H45N2O9PRS',
+    '3oocoa': 'C29H44N7O18P3S',
+    '3ooctACP': 'C19H33N2O9PRS',
+    '3ooctdACP': 'C29H53N2O9PRS',
+    '3oodcoa': 'C39H64N7O18P3S',
+    '3opalmACP': 'C27H49N2O9PRS',
+    '3ophb': 'C47H69O3',
+    '3otdcoa': 'C35H56N7O18P3S',
+    '3oxdhscoa': 'C29H39N7O20P3S',
+    '3pg': 'C3H4O7P',
+    '3php': 'C3H2O7P',
+    '3psme': 'C10H9O10P',
+    '3sala': 'C3H5NO4S',
+    '3ump_e': 'C9H11N2O9P',
+    '3ump_p': 'C9H11N2O9P',
+    '4abut': 'C4H9NO2',
+    '4abut_e': 'C4H9NO2',
+    '4abut_p': 'C4H9NO2',
+    '4abutn': 'C4H10NO',
+    '4abz': 'C7H6NO2',
+    '4adcho': 'C10H10NO5',
+    '4ahmmp': 'C6H9N3O',
+    '4ampm': 'C6H8N3O4P',
+    '4c2me': 'C14H23N3O14P2',
+    '4crsol': 'C7H8O',
+    '4fe4s': 'S4Fe4',
+    '4h2opntn': 'C5H7O4',
+    '4hbz': 'C7H5O3',
+    '4hoxpacd_e': 'C8H8O2',
+    '4hoxpacd_p': 'C8H8O2',
+    '4hthr': 'C4H9NO4',
+    '4mhetz': 'C6H9NOS',
+    '4mop': 'C6H9O3',
+    '4mpetz': 'C6H8NO4PS',
+    '4pasp': 'C4H6NO7P',
+    '4per': 'C4H6O8P',
+    '4ppan': 'C9H15NO8P',
+    '4ppcys': 'C12H20N2O9PS',
+    '4r5au': 'C9H16N4O6',
+    '56dura': 'C4H6N2O2',
+    '5aizc': 'C9H11N3O9P',
+    '5aop': 'C5H9NO3',
+    '5aprbu': 'C9H15N4O9P',
+    '5apru': 'C9H13N4O9P',
+    '5caiz': 'C9H11N3O9P',
+    '5dglcn': 'C6H9O7',
+    '5dglcn_e': 'C6H9O7',
+    '5dglcn_p': 'C6H9O7',
+    '5dh4dglc': 'C6H6O7',
+    '5drib': 'C5H10O4',
+    '5fthf': 'C20H21N7O7',
+    '5mta': 'C11H15N5O3S',
+    '5mthf': 'C20H24N7O6',
+    '5mtr': 'C6H12O4S',
+    '5mtr_e': 'C6H12O4S',
+    '5mtr_p': 'C6H12O4S',
+    '5prdmbz': 'C14H17N2O7P',
+    '6hmhpt': 'C7H9N5O2',
+    '6hmhptpp': 'C7H8N5O8P2',
+    '6pgc': 'C6H10O10P',
+    '6pgl': 'C6H9O9P',
+    '8aonn': 'C9H17NO3',
+    'aacald': 'C2H6NO',
+    'aacoa': 'C25H36N7O18P3S',
+    'aact': 'C3H8NO',
+    'ac': 'C2H3O2',
+    'ac_e': 'C2H3O2',
+    'ac_p': 'C2H3O2',
+    'acac': 'C4H5O3',
+    'acac_e': 'C4H5O3',
+    'acac_p': 'C4H5O3',
+    'acACP': 'C13H23N2O8PRS',
+    'acald': 'C2H4O',
+    'acald_e': 'C2H4O',
+    'acald_p': 'C2H4O',
+    'acanth': 'C9H8NO3',
+    'accoa': 'C23H34N7O17P3S',
+    'acetol': 'C3H6O2',
+    'acetone': 'C3H6O',
+    'acetone_e': 'C3H6O',
+    'acg5p': 'C7H9NO8P',
+    'acg5sa': 'C7H10NO4',
+    'acgal1p_e': 'C8H14NO9P',
+    'acgal1p_p': 'C8H14NO9P',
+    'acgal_e': 'C8H15NO6',
+    'acgal_p': 'C8H15NO6',
+    'acgam': 'C8H15NO6',
+    'acgam1p': 'C8H14NO9P',
+    'acgam1p_e': 'C8H14NO9P',
+    'acgam1p_p': 'C8H14NO9P',
+    'acgam6p': 'C8H14NO9P',
+    'acgam_e': 'C8H15NO6',
+    'acgam_p': 'C8H15NO6',
+    'acglc_D': 'C8H14O7',
+    'acglu': 'C7H9NO5',
+    'acmalt': 'C14H24O12',
+    'acmana': 'C8H15NO6',
+    'acmana_e': 'C8H15NO6',
+    'acmana_p': 'C8H15NO6',
+    'acmanap': 'C8H14NO9P',
+    'acmum6p': 'C11H17NO11P',
+    'acmum_e': 'C11H18NO8',
+    'acmum_p': 'C11H18NO8',
+    'acnam': 'C11H18NO9',
+    'acnam_e': 'C11H18NO9',
+    'acnam_p': 'C11H18NO9',
+    'acolipa_e': 'C181H314N3O103P4',
+    'acolipa_p': 'C181H314N3O103P4',
+    'acon_C': 'C6H3O6',
+    'acon_T': 'C6H3O6',
+    'aconm': 'C7H6O6',
+    'acorn': 'C7H14N2O3',
+    'ACP': 'C11H21N2O7PRS',
+    'acser': 'C5H9NO4',
+    'acser_e': 'C5H9NO4',
+    'acser_p': 'C5H9NO4',
+    'actACP': 'C15H25N2O9PRS',
+    'actp': 'C2H3O5P',
+    'ade': 'C5H5N5',
+    'ade_e': 'C5H5N5',
+    'ade_p': 'C5H5N5',
+    'adn': 'C10H13N5O4',
+    'adn_e': 'C10H13N5O4',
+    'adn_p': 'C10H13N5O4',
+    'adocbi': 'C58H84CoN16O11',
+    'adocbip': 'C58H83CoN16O14P',
+    'adocbl': 'C72H100CoN18O17P',
+    'adocbl_e': 'C72H100CoN18O17P',
+    'adocbl_p': 'C72H100CoN18O17P',
+    'adp': 'C10H12N5O10P2',
+    'adpglc': 'C16H23N5O15P2',
+    'adphep_DD': 'C17H25N5O16P2',
+    'adphep_LD': 'C17H25N5O16P2',
+    'adprib': 'C15H21N5O14P2',
+    'ag': 'Ag',
+    'ag_e': 'Ag',
+    'agdpcbi': 'C68H95CoN21O21P2',
+    'agm': 'C5H16N4',
+    'agm_e': 'C5H16N4',
+    'agm_p': 'C5H16N4',
+    'ahcys': 'C14H20N6O5S',
+    'ahdt': 'C9H12N5O13P3',
+    'aicar': 'C9H13N4O8P',
+    'air': 'C8H12N3O7P',
+    'akg': 'C5H4O5',
+    'akg_e': 'C5H4O5',
+    'akg_p': 'C5H4O5',
+    'ala_B': 'C3H7NO2',
+    'ala_B_e': 'C3H7NO2',
+    'ala_B_p': 'C3H7NO2',
+    'ala_D': 'C3H7NO2',
+    'ala_D_e': 'C3H7NO2',
+    'ala_D_p': 'C3H7NO2',
+    'ala_L': 'C3H7NO2',
+    'ala_L_e': 'C3H7NO2',
+    'ala_L_p': 'C3H7NO2',
+    'alaala': 'C6H12N2O3',
+    'alaala_e': 'C6H12N2O3',
+    'alaala_p': 'C6H12N2O3',
+    'alac_S': 'C5H7O4',
+    'alatrna': 'C3H6NOR',
+    'all6p': 'C6H11O9P',
+    'all_D': 'C6H12O6',
+    'all_D_e': 'C6H12O6',
+    'all_D_p': 'C6H12O6',
+    'alltn': 'C4H6N4O3',
+    'alltn_e': 'C4H6N4O3',
+    'alltn_p': 'C4H6N4O3',
+    'alltt': 'C4H7N4O4',
+    'allul6p': 'C6H11O9P',
+    'alpp_p': 'X',
+    'altrn': 'C6H11O7',
+    'amet': 'C15H23N6O5S',
+    'ametam': 'C14H24N6O3S',
+    'amob': 'C15H19N5O6S',
+    'amp': 'C10H12N5O7P',
+    'amp_e': 'C10H12N5O7P',
+    'amp_p': 'C10H12N5O7P',
+    'anhgm': 'C19H29N2O12',
+    'anhgm3p': 'C34H52N6O19',
+    'anhgm3p_p': 'C34H52N6O19',
+    'anhgm4p': 'C37H57N7O20',
+    'anhgm4p_p': 'C37H57N7O20',
+    'anhgm_e': 'C19H29N2O12',
+    'anhgm_p': 'C19H29N2O12',
+    'anhm': 'C11H16NO7',
+    'anhm3p': 'C26H39N5O14',
+    'anhm4p': 'C29H44N6O15',
+    'anth': 'C7H6NO2',
+    'ap4a': 'C20H24N10O19P4',
+    'ap5a': 'C20H24N10O22P5',
+    'apg120': 'C42H80O11P',
+    'apg140': 'C48H92O11P',
+    'apg141': 'C48H86O11P',
+    'apg160': 'C54H104O11P',
+    'apg161': 'C54H98O11P',
+    'apg180': 'C60H116O11P',
+    'apg181': 'C60H110O11P',
+    'apoACP': 'RHO',
+    'appl': 'C3H10NO',
+    'aps': 'C10H12N5O10PS',
+    'ara5p': 'C5H9O8P',
+    'arab_L': 'C5H10O5',
+    'arab_L_e': 'C5H10O5',
+    'arab_L_p': 'C5H10O5',
+    'aragund': 'C71H115N1O17P2',
+    'arbt6p': 'C12H15O10P',
+    'arbt_e': 'C12H16O7',
+    'arbt_p': 'C12H16O7',
+    'arbtn': 'C22H33N4O13',
+    'arbtn_e': 'C22H33N4O13',
+    'arbtn_fe3': 'C22H33FeN4O13',
+    'arbtn_fe3_e': 'C22H33FeN4O13',
+    'arbtn_fe3_p': 'C22H33FeN4O13',
+    'arbtn_p': 'C22H33N4O13',
+    'arg_L': 'C6H15N4O2',
+    'arg_L_e': 'C6H15N4O2',
+    'arg_L_p': 'C6H15N4O2',
+    'argsuc': 'C10H17N4O6',
+    'argtrna': 'C6H14N4OR',
+    'ascb6p': 'C6H7O9P',
+    'ascb_L_e': 'C6H8O6',
+    'ascb_L_p': 'C6H8O6',
+    'asn_L': 'C4H8N2O3',
+    'asn_L_e': 'C4H8N2O3',
+    'asn_L_p': 'C4H8N2O3',
+    'asntrna': 'C14H24N2O12PR2',
+    'aso3': 'AsO3',
+    'aso3_e': 'AsO3',
+    'aso3_p': 'AsO3',
+    'aso4': 'AsO4',
+    'asp_L': 'C4H6NO4',
+    'asp_L_e': 'C4H6NO4',
+    'asp_L_p': 'C4H6NO4',
+    'aspsa': 'C4H7NO3',
+    'asptrna': 'C4H5NO3R',
+    'athr_L': 'C4H9NO3',
+    'athtp': 'C22H28N9O13P3S',
+    'atp': 'C10H12N5O13P3',
+    'b2coa': 'C25H36N7O17P3S',
+    'bbtcoa': 'C28H46N8O17P3S',
+    'betald': 'C5H12NO',
+    'bglycogen': 'C6H10O5',
+    'bmoco': 'C20H20N10O13P2S4Mo',
+    'bmoco1gdp': 'C30H32N15O20P3S4Mo',
+    'bmocogdp': 'C40H44N20O27P4S4Mo',
+    'btal': 'C4H8O',
+    'btcoa': 'C25H38N7O17P3S',
+    'btn': 'C10H15N2O3S',
+    'btn_e': 'C10H15N2O3S',
+    'btn_p': 'C10H15N2O3S',
+    'btnso': 'C10H15N2O4S',
+    'but': 'C4H7O2',
+    'but2eACP': 'C15H25N2O8PRS',
+    'but_e': 'C4H7O2',
+    'but_p': 'C4H7O2',
+    'butACP': 'C15H27N2O8PRS',
+    'butso3': 'C4H9O3S',
+    'butso3_e': 'C4H9O3S',
+    'butso3_p': 'C4H9O3S',
+    'bwco': 'C20H20N10O13P2S4W',
+    'bwco1gdp': 'C30H32N15O20P3S4W',
+    'bwcogdp': 'C40H44N20O27P4S4W',
+    'ca2': 'Ca',
+    'ca2_e': 'Ca',
+    'ca2_p': 'Ca',
+    'camp': 'C10H11N5O6P',
+    'cbasp': 'C5H6N2O5',
+    'cbi': 'C48H72CoN11O8',
+    'cbi_e': 'C48H72CoN11O8',
+    'cbi_p': 'C48H72CoN11O8',
+    'cbl1': 'C62H88CoN13O14P',
+    'cbl1_e': 'C62H88CoN13O14P',
+    'cbl1_p': 'C62H88CoN13O14P',
+    'cbm': 'CH2NO2',
+    'cbp': 'CH2NO5P',
+    'cd2': 'Cd',
+    'cd2_e': 'Cd',
+    'cd2_p': 'Cd',
+    'cddec5eACP': 'C23H41N2O8PRS',
+    'cdec3eACP': 'C21H37N2O8PRS',
+    'cdg': 'C7H5N4O3',
+    'cdp': 'C9H12N3O11P2',
+    'cdpdddecg': 'C36H63N3O15P2',
+    'cdpdhdec9eg': 'C44H75N3O15P2',
+    'cdpdhdecg': 'C44H79N3O15P2',
+    'cdpdodec11eg': 'C48H83N3O15P2',
+    'cdpdodecg': 'C48H87N3O15P2',
+    'cdpdtdec7eg': 'C40H67N3O15P2',
+    'cdpdtdecg': 'C40H71N3O15P2',
+    'cechddd': 'C9H11O4',
+    'cenchddd': 'C9H9O4',
+    'cgly': 'C5H10N2O3S',
+    'cgly_e': 'C5H10N2O3S',
+    'cgly_p': 'C5H10N2O3S',
+    'chol': 'C5H14NO',
+    'chol_e': 'C5H14NO',
+    'chol_p': 'C5H14NO',
+    'chor': 'C10H8O6',
+    'chtbs6p': 'C16H27N2O14P',
+    'chtbs_e': 'C16H28N2O11',
+    'chtbs_p': 'C16H28N2O11',
+    'cinnm': 'C9H7O2',
+    'cit': 'C6H5O7',
+    'cit_e': 'C6H5O7',
+    'cit_p': 'C6H5O7',
+    'citr_L': 'C6H13N3O3',
+    'ckdo': 'C17H24N3O15P',
+    'cl': 'Cl',
+    'cl_e': 'Cl',
+    'cl_p': 'Cl',
+    'clpn120_p': 'C57H108O17P2',
+    'clpn140_p': 'C65H124O17P2',
+    'clpn141_p': 'C65H116O17P2',
+    'clpn160_p': 'C73H140O17P2',
+    'clpn161_p': 'C73H132O17P2',
+    'clpn180_p': 'C81H156O17P2',
+    'clpn181_p': 'C81H148O17P2',
+    'cm_e': 'C11H12Cl2N2O5',
+    'cm_p': 'C11H12Cl2N2O5',
+    'cmp': 'C9H12N3O8P',
+    'cmp_e': 'C9H12N3O8P',
+    'cmp_p': 'C9H12N3O8P',
+    'co2': 'CO2',
+    'co2_e': 'CO2',
+    'co2_p': 'CO2',
+    'coa': 'C21H32N7O16P3S',
+    'cobalt2': 'Co',
+    'cobalt2_e': 'Co',
+    'cobalt2_p': 'Co',
+    'colipa': 'C176H303N2O100P4',
+    'colipa_e': 'C176H303N2O100P4',
+    'colipa_p': 'C176H303N2O100P4',
+    'colipap_e': 'C176H303N2O103P5',
+    'colipap_p': 'C176H303N2O103P5',
+    'cpe160': 'C39H74N1O8P1',
+    'cpe180': 'C43H82N1O8P1',
+    'cpg160': 'C40H74O10P1',
+    'cpg180': 'C44H82O10P1',
+    'cpgn': 'C35H52N6O13Fe',
+    'cpgn_e': 'C35H52N6O13Fe',
+    'cpgn_p': 'C35H52N6O13Fe',
+    'cpgn_un': 'C35H52N6O13',
+    'cpgn_un_e': 'C35H52N6O13',
+    'cpgn_un_p': 'C35H52N6O13',
+    'cph4': 'C7H8N5O3',
+    'cpmp': 'C10H13N5O8P',
+    'cpppg3': 'C36H40N4O8',
+    'crn': 'C7H15NO3',
+    'crn_D': 'C7H15NO3',
+    'crn_D_e': 'C7H15NO3',
+    'crn_D_p': 'C7H15NO3',
+    'crn_e': 'C7H15NO3',
+    'crn_p': 'C7H15NO3',
+    'crncoa': 'C28H46N8O18P3S',
+    'crnDcoa': 'C28H46N8O18P3S',
+    'csn': 'C4H5N3O',
+    'csn_e': 'C4H5N3O',
+    'csn_p': 'C4H5N3O',
+    'ctbt': 'C7H13NO2',
+    'ctbt_p': 'C7H13NO2',
+    'ctbtcoa': 'C28H44N8O17P3S',
+    'ctp': 'C9H12N3O14P3',
+    'cu': 'Cu',
+    'cu2': 'Cu',
+    'cu2_e': 'Cu',
+    'cu2_p': 'Cu',
+    'cu_e': 'Cu',
+    'cu_p': 'Cu',
+    'cyan': 'CHN',
+    'cyan_e': 'CHN',
+    'cyan_p': 'CHN',
+    'cynt': 'CNO',
+    'cynt_e': 'CNO',
+    'cynt_p': 'CNO',
+    'cys_D': 'C3H7NO2S',
+    'cys_D_e': 'C3H7NO2S',
+    'cys_D_p': 'C3H7NO2S',
+    'cys_L': 'C3H7NO2S',
+    'cys_L_e': 'C3H7NO2S',
+    'cys_L_p': 'C3H7NO2S',
+    'cyst_L': 'C7H14N2O4S',
+    'cystrna': 'C3H6NOSR',
+    'cytd': 'C9H13N3O5',
+    'cytd_e': 'C9H13N3O5',
+    'cytd_p': 'C9H13N3O5',
+    'dad_2': 'C10H13N5O3',
+    'dad_2_e': 'C10H13N5O3',
+    'dad_2_p': 'C10H13N5O3',
+    'dad_5': 'C10H13N5O3',
+    'dadp': 'C10H12N5O9P2',
+    'damp': 'C10H12N5O6P',
+    'damp_e': 'C10H12N5O6P',
+    'damp_p': 'C10H12N5O6P',
+    'dann': 'C9H21N2O2',
+    'datp': 'C10H12N5O12P3',
+    'db4p': 'C4H7O6P',
+    'dc2coa': 'C31H48N7O17P3S',
+    'dca': 'C10H19O2',
+    'dca_e': 'C10H19O2',
+    'dca_p': 'C10H19O2',
+    'dcaACP': 'C21H39N2O8PRS',
+    'dcacoa': 'C31H50N7O17P3S',
+    'dcamp': 'C14H14N5O11P',
+    'dcdp': 'C9H12N3O10P2',
+    'dcmp': 'C9H12N3O7P',
+    'dcmp_e': 'C9H12N3O7P',
+    'dcmp_p': 'C9H12N3O7P',
+    'dctp': 'C9H12N3O13P3',
+    'dcyt': 'C9H13N3O4',
+    'dcyt_e': 'C9H13N3O4',
+    'dcyt_p': 'C9H13N3O4',
+    'dd2coa': 'C33H52N7O17P3S',
+    'ddca': 'C12H23O2',
+    'ddca_e': 'C12H23O2',
+    'ddca_p': 'C12H23O2',
+    'ddcaACP': 'C23H43N2O8PRS',
+    'ddcacoa': 'C33H54N7O17P3S',
+    'ddcap': 'C12H24O5P1',
+    'dgdp': 'C10H12N5O10P2',
+    'dgmp': 'C10H12N5O7P',
+    'dgmp_e': 'C10H12N5O7P',
+    'dgmp_p': 'C10H12N5O7P',
+    'dgslnt': 'C20H30N6O12S2Se',
+    'dgsn': 'C10H13N5O4',
+    'dgsn_e': 'C10H13N5O4',
+    'dgsn_p': 'C10H13N5O4',
+    'dgtp': 'C10H12N5O13P3',
+    'dha': 'C3H6O3',
+    'dha_e': 'C3H6O3',
+    'dha_p': 'C3H6O3',
+    'dhap': 'C3H5O6P',
+    'dhcinnm': 'C9H7O4',
+    'dhf': 'C19H19N7O6',
+    'dhgly': 'C2H2NO2',
+    'dhmpt': 'C9H13N5O4',
+    'dhmptp': 'C9H12N5O13P3',
+    'dhna': 'C11H7O4',
+    'dhnpt': 'C9H13N5O4',
+    'dhor_S': 'C5H5N2O4',
+    'dhpmp': 'C9H12N5O7P',
+    'dhpppn': 'C9H9O4',
+    'dhpt': 'C14H13N6O3',
+    'dhptd': 'C5H8O4',
+    'dhptdn': 'C6H6N4',
+    'didp': 'C10H11N4O10P2',
+    'dimp': 'C10H11N4O7P',
+    'dimp_e': 'C10H11N4O7P',
+    'dimp_p': 'C10H11N4O7P',
+    'din': 'C10H12N4O4',
+    'din_e': 'C10H12N4O4',
+    'din_p': 'C10H12N4O4',
+    'ditp': 'C10H11N4O13P3',
+    'dmbzid': 'C9H10N2',
+    'dmlz': 'C13H18N4O6',
+    'dmpp': 'C5H9O7P2',
+    'dms': 'C2H6S',
+    'dms_e': 'C2H6S',
+    'dms_p': 'C2H6S',
+    'dmso': 'C2H6OS',
+    'dmso_e': 'C2H6OS',
+    'dmso_p': 'C2H6OS',
+    'dnad': 'C21H24N6O15P2',
+    'dopa_e': 'C8H12NO2',
+    'dopa_p': 'C8H12NO2',
+    'doxrbcn_e': 'C27H29NO11',
+    'doxrbcn_p': 'C27H29NO11',
+    'dpcoa': 'C21H33N7O13P2S',
+    'dsbaox_p': 'X',
+    'dsbard_p': 'XH2',
+    'dsbcox_p': 'X',
+    'dsbcrd_p': 'XH2',
+    'dsbdox': 'X',
+    'dsbdrd': 'XH2',
+    'dsbgox_p': 'X',
+    'dsbgrd_p': 'XH2',
+    'dscl': 'C42H41N4O16',
+    'dtbt': 'C10H17N2O3',
+    'dtdp': 'C10H13N2O11P2',
+    'dtdp4aaddg': 'C18H27N3O15P2',
+    'dtdp4addg': 'C16H26N3O14P2',
+    'dtdp4d6dg': 'C16H22N2O15P2',
+    'dtdp4d6dm': 'C16H22N2O15P2',
+    'dtdpglu': 'C16H24N2O16P2',
+    'dtdprmn': 'C16H24N2O15P2',
+    'dtmp': 'C10H13N2O8P',
+    'dtmp_e': 'C10H13N2O8P',
+    'dtmp_p': 'C10H13N2O8P',
+    'dttp': 'C10H13N2O14P3',
+    'dudp': 'C9H11N2O11P2',
+    'dump': 'C9H11N2O8P',
+    'dump_e': 'C9H11N2O8P',
+    'dump_p': 'C9H11N2O8P',
+    'duri': 'C9H12N2O5',
+    'duri_e': 'C9H12N2O5',
+    'duri_p': 'C9H12N2O5',
+    'dutp': 'C9H11N2O14P3',
+    'dxyl': 'C5H10O4',
+    'dxyl5p': 'C5H9O7P',
+    'e4p': 'C4H7O7P',
+    'eca2und_p': 'C103H162N6O37P2',
+    'eca3und_p': 'C127H198N9O52P2',
+    'eca4colipa_e': 'C272H447N14O160P4',
+    'eca4colipa_p': 'C272H447N14O160P4',
+    'eca4und_p': 'C151H234N12O67P2',
+    'egmeACP': 'C17H27N2O10PRS',
+    'eig3p': 'C6H9N2O6P',
+    'enlipa_e': 'C112H202N3O42P3',
+    'enlipa_p': 'C112H202N3O42P3',
+    'enter': 'C30H27N3O15',
+    'enter_e': 'C30H27N3O15',
+    'enter_p': 'C30H27N3O15',
+    'epmeACP': 'C19H31N2O10PRS',
+    'etha': 'C2H8NO',
+    'etha_e': 'C2H8NO',
+    'etha_p': 'C2H8NO',
+    'ethso3': 'C2H5O3S',
+    'ethso3_e': 'C2H5O3S',
+    'ethso3_p': 'C2H5O3S',
+    'etoh': 'C2H6O',
+    'etoh_e': 'C2H6O',
+    'etoh_p': 'C2H6O',
+    'f1p': 'C6H11O9P',
+    'f6p': 'C6H11O9P',
+    'f6p_e': 'C6H11O9P',
+    'f6p_p': 'C6H11O9P',
+    'fad': 'C27H31N9O15P2',
+    'fadh2': 'C27H33N9O15P2',
+    'fald': 'CH2O',
+    'fald_e': 'CH2O',
+    'fald_p': 'CH2O',
+    'fc1p': 'C6H11O8P',
+    'fcl_L': 'C6H12O5',
+    'fdp': 'C6H10O12P2',
+    'fe2': 'Fe',
+    'fe2_e': 'Fe',
+    'fe2_p': 'Fe',
+    'fe3': 'Fe',
+    'fe3_e': 'Fe',
+    'fe3_p': 'Fe',
+    'fe3dcit_e': 'C12H10FeO14',
+    'fe3dcit_p': 'C12H10FeO14',
+    'fe3dhbzs': 'C10H10NO6Fe',
+    'fe3dhbzs_e': 'C10H10NO6Fe',
+    'fe3dhbzs_p': 'C10H10NO6Fe',
+    'fe3hox': 'C9H18O6N3Fe',
+    'fe3hox_e': 'C9H18O6N3Fe',
+    'fe3hox_p': 'C9H18O6N3Fe',
+    'fe3hox_un': 'C9H18O6N3',
+    'fe3hox_un_e': 'C9H18O6N3',
+    'fe3hox_un_p': 'C9H18O6N3',
+    'fecrm': 'C27H42FeN9O12',
+    'fecrm_e': 'C27H42FeN9O12',
+    'fecrm_p': 'C27H42FeN9O12',
+    'fecrm_un': 'C27H42N9O12',
+    'fecrm_un_e': 'C27H42N9O12',
+    'fecrm_un_p': 'C27H42N9O12',
+    'feenter': 'C30H27FeN3O15',
+    'feenter_e': 'C30H27FeN3O15',
+    'feenter_p': 'C30H27FeN3O15',
+    'feoxam': 'C25H46FeN6O8',
+    'feoxam_e': 'C25H46FeN6O8',
+    'feoxam_p': 'C25H46FeN6O8',
+    'feoxam_un': 'C25H46N6O8',
+    'feoxam_un_e': 'C25H46N6O8',
+    'feoxam_un_p': 'C25H46N6O8',
+    'fgam': 'C8H13N2O9P',
+    'flxr': 'XH',
+    'flxso': 'XH',
+    'fmettrna': 'C6H9NO2SR',
+    'fmn': 'C17H19N4O9P',
+    'fmnh2': 'C17H21N4O9P',
+    'for': 'CH1O2',
+    'for_e': 'CH1O2',
+    'for_p': 'CH1O2',
+    'forcoa': 'C22H32N7O17P3S',
+    'fpram': 'C8H15N3O8P',
+    'fprica': 'C10H13N4O9P',
+    'frdp': 'C15H25O7P2',
+    'fru': 'C6H12O6',
+    'fru_e': 'C6H12O6',
+    'fru_p': 'C6H12O6',
+    'frulys': 'C12H25N2O7',
+    'frulys_e': 'C12H25N2O7',
+    'frulys_p': 'C12H25N2O7',
+    'frulysp': 'C12H24N2O10P',
+    'fruur': 'C6H9O7',
+    'fruur_e': 'C6H9O7',
+    'fruur_p': 'C6H9O7',
+    'fuc_L': 'C6H12O5',
+    'fuc_L_e': 'C6H12O5',
+    'fuc_L_p': 'C6H12O5',
+    'fum': 'C4H2O4',
+    'fum_e': 'C4H2O4',
+    'fum_p': 'C4H2O4',
+    'fusa_e': 'C31H48O5',
+    'fusa_p': 'C31H48O5',
+    'g1p': 'C6H11O9P',
+    'g1p_e': 'C6H11O9P',
+    'g1p_p': 'C6H11O9P',
+    'g3p': 'C3H5O6P',
+    'g3pc': 'C8H20NO6P',
+    'g3pc_e': 'C8H20NO6P',
+    'g3pc_p': 'C8H20NO6P',
+    'g3pe': 'C5H14NO6P',
+    'g3pe_e': 'C5H14NO6P',
+    'g3pe_p': 'C5H14NO6P',
+    'g3pg': 'C6H14O8P',
+    'g3pg_e': 'C6H14O8P',
+    'g3pg_p': 'C6H14O8P',
+    'g3pi': 'C9H18O11P',
+    'g3pi_e': 'C9H18O11P',
+    'g3pi_p': 'C9H18O11P',
+    'g3ps': 'C6H13NO8P',
+    'g3ps_e': 'C6H13NO8P',
+    'g3ps_p': 'C6H13NO8P',
+    'g6p': 'C6H11O9P',
+    'g6p_e': 'C6H11O9P',
+    'g6p_p': 'C6H11O9P',
+    'gagicolipa': 'C157H271N2O84P4',
+    'gal': 'C6H12O6',
+    'gal1p': 'C6H11O9P',
+    'gal1p_e': 'C6H11O9P',
+    'gal1p_p': 'C6H11O9P',
+    'gal_bD_e': 'C6H12O6',
+    'gal_bD_p': 'C6H12O6',
+    'gal_e': 'C6H12O6',
+    'gal_p': 'C6H12O6',
+    'galct_D': 'C6H8O8',
+    'galct_D_e': 'C6H8O8',
+    'galct_D_p': 'C6H8O8',
+    'galctn_D': 'C6H11O7',
+    'galctn_D_e': 'C6H11O7',
+    'galctn_D_p': 'C6H11O7',
+    'galctn_L': 'C6H11O7',
+    'galctn_L_e': 'C6H11O7',
+    'galctn_L_p': 'C6H11O7',
+    'galt1p': 'C6H13O9P',
+    'galt_e': 'C6H14O6',
+    'galt_p': 'C6H14O6',
+    'galur': 'C6H9O7',
+    'galur_e': 'C6H9O7',
+    'galur_p': 'C6H9O7',
+    'gam1p': 'C6H13NO8P',
+    'gam6p': 'C6H13NO8P',
+    'gam6p_e': 'C6H13NO8P',
+    'gam6p_p': 'C6H13NO8P',
+    'gam_e': 'C6H14NO5',
+    'gam_p': 'C6H14NO5',
+    'gar': 'C7H14N2O8P',
+    'garagund': 'C77H125N1O22P2',
+    'gbbtn': 'C7H15NO2',
+    'gbbtn_e': 'C7H15NO2',
+    'gbbtn_p': 'C7H15NO2',
+    'gcald': 'C2H4O2',
+    'gdp': 'C10H12N5O11P2',
+    'gdp_e': 'C10H12N5O11P2',
+    'gdp_p': 'C10H12N5O11P2',
+    'gdpddman': 'C16H21N5O15P2',
+    'gdpfuc': 'C16H23N5O15P2',
+    'gdpmann': 'C16H23N5O16P2',
+    'gdpofuc': 'C16H21N5O15P2',
+    'gdptp': 'C10H11N5O20P5',
+    'gfgaragund': 'C83H135N1O27P2',
+    'gg4abut': 'C9H15O5N2',
+    'ggagicolipa': 'C163H281N2O89P4',
+    'ggbutal': 'C9H16O4N2',
+    'gggagicolipa': 'C169H291N2O94P4',
+    'ggptrc': 'C9H20O3N3',
+    'ghb': 'C4H7O3',
+    'gicolipa': 'C151H261N2O79P4',
+    'glc_D': 'C6H12O6',
+    'glc_D_e': 'C6H12O6',
+    'glc_D_p': 'C6H12O6',
+    'glcn': 'C6H11O7',
+    'glcn_e': 'C6H11O7',
+    'glcn_p': 'C6H11O7',
+    'glcr': 'C6H8O8',
+    'glcr_e': 'C6H8O8',
+    'glcr_p': 'C6H8O8',
+    'glcur': 'C6H9O7',
+    'glcur1p_e': 'C6H8O10P',
+    'glcur1p_p': 'C6H8O10P',
+    'glcur_e': 'C6H9O7',
+    'glcur_p': 'C6H9O7',
+    'gln_L': 'C5H10N2O3',
+    'gln_L_e': 'C5H10N2O3',
+    'gln_L_p': 'C5H10N2O3',
+    'glntrna': 'C5H9N2O2R',
+    'glu1sa': 'C5H9NO3',
+    'glu5p': 'C5H8NO7P',
+    'glu5sa': 'C5H9NO3',
+    'glu_D': 'C5H8NO4',
+    'glu_L': 'C5H8NO4',
+    'glu_L_e': 'C5H8NO4',
+    'glu_L_p': 'C5H8NO4',
+    'glucys': 'C8H13N2O5S',
+    'glutrna': 'C5H7NO3R',
+    'glx': 'C2H1O3',
+    'gly': 'C2H5NO2',
+    'gly_e': 'C2H5NO2',
+    'gly_p': 'C2H5NO2',
+    'glyald': 'C3H6O3',
+    'glyald_e': 'C3H6O3',
+    'glyald_p': 'C3H6O3',
+    'glyb': 'C5H11NO2',
+    'glyb_e': 'C5H11NO2',
+    'glyb_p': 'C5H11NO2',
+    'glyc': 'C3H8O3',
+    'glyc2p': 'C3H7O6P',
+    'glyc2p_e': 'C3H7O6P',
+    'glyc2p_p': 'C3H7O6P',
+    'glyc3p': 'C3H7O6P',
+    'glyc3p_e': 'C3H7O6P',
+    'glyc3p_p': 'C3H7O6P',
+    'glyc_e': 'C3H8O3',
+    'glyc_p': 'C3H8O3',
+    'glyc_R': 'C3H5O4',
+    'glyc_R_e': 'C3H5O4',
+    'glyc_R_p': 'C3H5O4',
+    'glyclt': 'C2H3O3',
+    'glyclt_e': 'C2H3O3',
+    'glyclt_p': 'C2H3O3',
+    'glycogen': 'C6H10O5',
+    'glytrna': 'C2H4NOR',
+    'gmeACP': 'C17H29N2O10PRS',
+    'gmhep17bp': 'C7H12O13P2',
+    'gmhep1p': 'C7H13O10P',
+    'gmhep7p': 'C7H13O10P',
+    'gmp': 'C10H12N5O8P',
+    'gmp_e': 'C10H12N5O8P',
+    'gmp_p': 'C10H12N5O8P',
+    'gp4g': 'C20H24N10O21P4',
+    'grdp': 'C10H17O7P2',
+    'grxox': 'X',
+    'grxrd': 'XH2',
+    'gslnt': 'C10H16N3O6SSe',
+    'gsn': 'C10H13N5O5',
+    'gsn_e': 'C10H13N5O5',
+    'gsn_p': 'C10H13N5O5',
+    'gthox': 'C20H30N6O12S2',
+    'gthox_e': 'C20H30N6O12S2',
+    'gthox_p': 'C20H30N6O12S2',
+    'gthrd': 'C10H16N3O6S',
+    'gthrd_e': 'C10H16N3O6S',
+    'gthrd_p': 'C10H16N3O6S',
+    'gtp': 'C10H12N5O14P3',
+    'gtp_e': 'C10H12N5O14P3',
+    'gtp_p': 'C10H12N5O14P3',
+    'gtspmd': 'C17H36N6O5S',
+    'gua': 'C5H5N5O',
+    'gua_e': 'C5H5N5O',
+    'gua_p': 'C5H5N5O',
+    'h': 'H',
+    'h2': 'H2',
+    'h2_e': 'H2',
+    'h2_p': 'H2',
+    'h2mb4p': 'C5H9O8P2',
+    'h2o': 'H2O',
+    'h2o2': 'H2O2',
+    'h2o2_e': 'H2O2',
+    'h2o2_p': 'H2O2',
+    'h2o_e': 'H2O',
+    'h2o_p': 'H2O',
+    'h2s': 'H2S',
+    'h2s_e': 'H2S',
+    'h2s_p': 'H2S',
+    'h_e': 'H',
+    'h_p': 'H',
+    'hacolipa_e': 'C192H333N2O101P4',
+    'halipa_e': 'C126H226N2O40P2',
+    'hco3': 'CHO3',
+    'hcys_L': 'C4H9NO2S',
+    'hdca': 'C16H31O2',
+    'hdca_e': 'C16H31O2',
+    'hdca_p': 'C16H31O2',
+    'hdcap': 'C16H32O5P',
+    'hdcea': 'C16H29O2',
+    'hdcea_e': 'C16H29O2',
+    'hdcea_p': 'C16H29O2',
+    'hdceap': 'C16H30O5P',
+    'hdcoa': 'C37H60N7O17P3S',
+    'hdd2coa': 'C37H60N7O17P3S',
+    'hdeACP': 'C27H49N2O8PRS',
+    'hemeO': 'C49H56FeN4O5',
+    'hexACP': 'C17H31N2O8PRS',
+    'hg2': 'Hg',
+    'hg2_e': 'Hg',
+    'hg2_p': 'Hg',
+    'hgmeACP': 'C17H29N2O11PRS',
+    'hhlipa': 'C124H220N2O51P2',
+    'his_L': 'C6H9N3O2',
+    'his_L_e': 'C6H9N3O2',
+    'his_L_p': 'C6H9N3O2',
+    'hisp': 'C6H11N3O4P',
+    'histd': 'C6H12N3O',
+    'histrna': 'C6H8N3OR',
+    'hkndd': 'C9H8O6',
+    'hkntd': 'C9H6O6',
+    'hlipa': 'C117H208N2O45P2',
+    'hmbil': 'C40H38N4O17',
+    'hmgth': 'C11H18N3O7S',
+    'hom_L': 'C4H9NO3',
+    'hom_L_e': 'C4H9NO3',
+    'hom_L_p': 'C4H9NO3',
+    'hphhlipa': 'C131H231N2O60P3',
+    'hpmeACP': 'C19H33N2O11PRS',
+    'hpyr': 'C3H3O4',
+    'hqn': 'C6H6O2',
+    'hx2coa': 'C27H40N7O17P3S',
+    'hxa': 'C6H11O2',
+    'hxa_e': 'C6H11O2',
+    'hxa_p': 'C6H11O2',
+    'hxan': 'C5H4N4O',
+    'hxan_e': 'C5H4N4O',
+    'hxan_p': 'C5H4N4O',
+    'hxcoa': 'C27H42N7O17P3S',
+    'iasp': 'C4H3NO4',
+    'ichor': 'C10H8O6',
+    'icit': 'C6H5O7',
+    'icolipa': 'C145H251N2O74P4',
+    'idon_L': 'C6H11O7',
+    'idon_L_e': 'C6H11O7',
+    'idon_L_p': 'C6H11O7',
+    'idp': 'C10H11N4O11P2',
+    'ile_L': 'C6H13NO2',
+    'ile_L_e': 'C6H13NO2',
+    'ile_L_p': 'C6H13NO2',
+    'iletrna': 'C6H12NOR',
+    'imacp': 'C6H7N2O5P',
+    'imp': 'C10H11N4O8P',
+    'imp_e': 'C10H11N4O8P',
+    'imp_p': 'C10H11N4O8P',
+    'indole': 'C8H7N',
+    'indole_e': 'C8H7N',
+    'indole_p': 'C8H7N',
+    'inost': 'C6H12O6',
+    'inost_e': 'C6H12O6',
+    'inost_p': 'C6H12O6',
+    'ins': 'C10H12N4O5',
+    'ins_e': 'C10H12N4O5',
+    'ins_p': 'C10H12N4O5',
+    'ipdp': 'C5H9O7P2',
+    'iscs': 'HSR',
+    'iscssh': 'HS2R',
+    'iscu': 'H8O2S6R',
+    'iscu_2fe2s': 'H4O2S8Fe2R',
+    'iscu_2fe2s2': 'O2S10Fe4R',
+    'iscu_4fe4s': 'H4O2S10Fe4R',
+    'isetac': 'C2H5O4S',
+    'isetac_e': 'C2H5O4S',
+    'isetac_p': 'C2H5O4S',
+    'itp': 'C10H11N4O14P3',
+    'k': 'K',
+    'k_e': 'K',
+    'k_p': 'K',
+    'kdo': 'C8H13O8',
+    'kdo2lipid4': 'C84H148N2O37P2',
+    'kdo2lipid4_e': 'C84H148N2O37P2',
+    'kdo2lipid4_p': 'C84H148N2O37P2',
+    'kdo2lipid4L': 'C96H170N2O38P2',
+    'kdo2lipid4p': 'C100H176N2O38P2',
+    'kdo8p': 'C8H12O11P',
+    'kdolipid4': 'C76H137N2O30P2',
+    'kphphhlipa': 'C139H241N2O70P4',
+    'lac_D': 'C3H5O3',
+    'lac_D_e': 'C3H5O3',
+    'lac_D_p': 'C3H5O3',
+    'lac_L': 'C3H5O3',
+    'lac_L_e': 'C3H5O3',
+    'lac_L_p': 'C3H5O3',
+    'LalaDglu': 'C8H13N2O5',
+    'LalaDglu_e': 'C8H13N2O5',
+    'LalaDglu_p': 'C8H13N2O5',
+    'LalaDgluMdap': 'C15H25N4O8',
+    'LalaDgluMdap_e': 'C15H25N4O8',
+    'LalaDgluMdap_p': 'C15H25N4O8',
+    'LalaDgluMdapDala': 'C18H30N5O9',
+    'LalaDgluMdapDala_e': 'C18H30N5O9',
+    'LalaDgluMdapDala_p': 'C18H30N5O9',
+    'LalaLglu': 'C8H13N2O5',
+    'LalaLglu_e': 'C8H13N2O5',
+    'LalaLglu_p': 'C8H13N2O5',
+    'lald_D': 'C3H6O2',
+    'lald_L': 'C3H6O2',
+    'lcts': 'C12H22O11',
+    'lcts_e': 'C12H22O11',
+    'lcts_p': 'C12H22O11',
+    'leu_L': 'C6H13NO2',
+    'leu_L_e': 'C6H13NO2',
+    'leu_L_p': 'C6H13NO2',
+    'leutrna': 'C6H12NOR',
+    'lgt_S': 'C13H20N3O8S',
+    'lipa': 'C110H196N2O39P2',
+    'lipa_cold_e': 'C114H202N2O39P2',
+    'lipa_e': 'C110H196N2O39P2',
+    'lipa_p': 'C110H196N2O39P2',
+    'lipaold': 'C114H202N2O39P2',
+    'lipaold_e': 'C114H202N2O39P2',
+    'lipaold_p': 'C114H202N2O39P2',
+    'lipidA': 'C68H126N2O23P2',
+    'lipidAds': 'C68H127N2O20P',
+    'lipidX': 'C34H64NO12P',
+    'lipoamp': 'C18H25N5O8PS2',
+    'lipoate': 'C8H14O2S2',
+    'lipoate_e': 'C8H14O2S2',
+    'lipoate_p': 'C8H14O2S2',
+    'lipopb': 'C8H13OS2',
+    'lpp_p': 'XC16H30O1',
+    'lys_L': 'C6H15N2O2',
+    'lys_L_e': 'C6H15N2O2',
+    'lys_L_p': 'C6H15N2O2',
+    'lystrna': 'C6H14N2OR',
+    'lyx_L': 'C5H10O5',
+    'lyx_L_e': 'C5H10O5',
+    'lyx_L_p': 'C5H10O5',
+    'mal_D': 'C4H4O5',
+    'mal_D_e': 'C4H4O5',
+    'mal_D_p': 'C4H4O5',
+    'mal_L': 'C4H4O5',
+    'mal_L_e': 'C4H4O5',
+    'mal_L_p': 'C4H4O5',
+    'malACP': 'C14H22N2O10PRS',
+    'malcoa': 'C24H33N7O19P3S',
+    'malcoame': 'C25H36N7O19P3S',
+    'malt': 'C12H22O11',
+    'malt6p': 'C12H21O14P',
+    'malt_e': 'C12H22O11',
+    'malt_p': 'C12H22O11',
+    'malthp': 'C42H72O36',
+    'malthx': 'C36H62O31',
+    'malthx_e': 'C36H62O31',
+    'malthx_p': 'C36H62O31',
+    'maltpt': 'C30H52O26',
+    'maltpt_e': 'C30H52O26',
+    'maltpt_p': 'C30H52O26',
+    'malttr': 'C18H32O16',
+    'malttr_e': 'C18H32O16',
+    'malttr_p': 'C18H32O16',
+    'maltttr': 'C24H42O21',
+    'maltttr_e': 'C24H42O21',
+    'maltttr_p': 'C24H42O21',
+    'man': 'C6H12O6',
+    'man1p': 'C6H11O9P',
+    'man6p': 'C6H11O9P',
+    'man6p_e': 'C6H11O9P',
+    'man6p_p': 'C6H11O9P',
+    'man6pglyc': 'C9H14O12P',
+    'man_e': 'C6H12O6',
+    'man_p': 'C6H12O6',
+    'mana': 'C6H11O7',
+    'manglyc_e': 'C9H15O9',
+    'manglyc_p': 'C9H15O9',
+    'mdhdhf': 'C5H8O4',
+    'melib': 'C12H22O11',
+    'melib_e': 'C12H22O11',
+    'melib_p': 'C12H22O11',
+    'meoh': 'CH4O1',
+    'meoh_e': 'CH4O1',
+    'meoh_p': 'CH4O1',
+    'mercppyr': 'C3H3O3S',
+    'met_D': 'C5H10NO2S',
+    'met_D_e': 'C5H10NO2S',
+    'met_D_p': 'C5H10NO2S',
+    'met_L': 'C5H11NO2S',
+    'met_L_e': 'C5H11NO2S',
+    'met_L_p': 'C5H11NO2S',
+    'methf': 'C20H20N7O6',
+    'metsox_R_L': 'C5H11NO3S',
+    'metsox_R_L_e': 'C5H11NO3S',
+    'metsox_R_L_p': 'C5H11NO3S',
+    'metsox_S_L': 'C5H11NO3S',
+    'metsox_S_L_e': 'C5H11NO3S',
+    'metsox_S_L_p': 'C5H11NO3S',
+    'mettrna': 'C5H10NOSR',
+    'mg2': 'Mg',
+    'mg2_e': 'Mg',
+    'mg2_p': 'Mg',
+    'mi1p_D': 'C6H11O9P',
+    'micit': 'C7H7O7',
+    'mincyc_e': 'C23H27N3O7',
+    'mincyc_p': 'C23H27N3O7',
+    'minohp_e': 'C6H6O24P6',
+    'minohp_p': 'C6H6O24P6',
+    'mlthf': 'C20H21N7O6',
+    'mmcoa_S': 'C25H35N7O19P3S',
+    'mmet': 'C6H14NO2S',
+    'mmet_e': 'C6H14NO2S',
+    'mmet_p': 'C6H14NO2S',
+    'mn2': 'Mn',
+    'mn2_e': 'Mn',
+    'mn2_p': 'Mn',
+    'mnl1p': 'C6H13O9P',
+    'mnl_e': 'C6H14O6',
+    'mnl_p': 'C6H14O6',
+    'moadamp': 'C11H12N5O8P1X',
+    'moadcoo': 'C1O2X',
+    'moadcosh': 'C1H1O1S1X',
+    'mobd': 'MoO4',
+    'mobd_e': 'MoO4',
+    'mobd_p': 'MoO4',
+    'moco': 'C10H10N5O8PS2Mo',
+    'mococdp': 'C19H22N8O15P2S2Mo',
+    'mocogdp': 'C20H22N10O15P2S2Mo',
+    'mpt': 'C10H10N5O6PS2Cu',
+    'mptamp': 'C20H22N10O12P2S2Cu',
+    'mql8': 'C51H74O2',
+    'mqn8': 'C51H72O2',
+    'msa': 'C3H3O3',
+    'mso3': 'CH3O3S',
+    'mso3_e': 'CH3O3S',
+    'mso3_p': 'CH3O3S',
+    'mthgxl': 'C3H4O2',
+    'mththf': 'C5H10O5',
+    'murein3p3p_p': 'C68H104N12O38',
+    'murein3px3p_p': 'C68H102N12O37',
+    'murein3px4p_p': 'C71H107N13O38',
+    'murein4p3p_p': 'C71H109N13O39',
+    'murein4p4p_p': 'C74H114N14O40',
+    'murein4px4p4p_p': 'C111H169N21O59',
+    'murein4px4p_p': 'C74H112N14O39',
+    'murein4px4px4p_p': 'C111H167N21O58',
+    'murein5p3p_p': 'C74H114N14O40',
+    'murein5p4p_p': 'C77H119N15O41',
+    'murein5p5p5p_p': 'C120H186N24O63',
+    'murein5p5p_p': 'C80H124N16O42',
+    'murein5px3p_p': 'C74H112N14O39',
+    'murein5px4p_p': 'C77H117N15O40',
+    'murein5px4px4p_p': 'C114H172N22O59',
+    'myrsACP': 'C25H47N2O8PRS',
+    'N1aspmd': 'C9H23N3O',
+    'n2o': 'N2O',
+    'n2o_e': 'N2O',
+    'n2o_p': 'N2O',
+    'n8aspmd': 'C9H23N3O',
+    'na1': 'Na',
+    'na15dap': 'C8H24N3',
+    'na1_e': 'Na',
+    'na1_p': 'Na',
+    'nac': 'C6H4NO2',
+    'nac_e': 'C6H4NO2',
+    'nac_p': 'C6H4NO2',
+    'nad': 'C21H26N7O14P2',
+    'nadh': 'C21H27N7O14P2',
+    'nadp': 'C21H25N7O17P3',
+    'nadph': 'C21H26N7O17P3',
+    'ncam': 'C6H6N2O',
+    'nh4': 'H4N',
+    'nh4_e': 'H4N',
+    'nh4_p': 'H4N',
+    'ni2': 'Ni',
+    'ni2_e': 'Ni',
+    'ni2_p': 'Ni',
+    'nicrnt': 'C11H12NO9P',
+    'nmn': 'C11H14N2O8P',
+    'nmn_e': 'C11H14N2O8P',
+    'nmn_p': 'C11H14N2O8P',
+    'Nmtrp': 'C12H14N2O2',
+    'no': 'NO',
+    'no2': 'NO2',
+    'no2_e': 'NO2',
+    'no2_p': 'NO2',
+    'no3': 'NO3',
+    'no3_e': 'NO3',
+    'no3_p': 'NO3',
+    'no_e': 'NO',
+    'no_p': 'NO',
+    'novbcn_e': 'C31H36N2O11',
+    'novbcn_p': 'C31H36N2O11',
+    'o16a2und_p': 'C123H200N2O57P2',
+    'o16a3und_p': 'C157H255N3O82P2',
+    'o16a4colipa_e': 'C312H523N6O200P4',
+    'o16a4colipa_p': 'C312H523N6O200P4',
+    'o16a4und_p': 'C191H310N4O107P2',
+    'o16aund': 'C89H145N1O32P2',
+    'o16aund_p': 'C89H145N1O32P2',
+    'o2': 'O2',
+    'o2_e': 'O2',
+    'o2_p': 'O2',
+    'o2s': 'O2',
+    'o2s_e': 'O2',
+    'o2s_p': 'O2',
+    'oaa': 'C4H2O5',
+    'oc2coa': 'C29H44N7O17P3S',
+    'ocACP': 'C19H35N2O8PRS',
+    'occoa': 'C29H46N7O17P3S',
+    'ocdca': 'C18H35O2',
+    'ocdca_e': 'C18H35O2',
+    'ocdca_p': 'C18H35O2',
+    'ocdcaACP': 'C29H55N2O8PRS',
+    'ocdcap': 'C18H36O5P',
+    'ocdcea': 'C18H33O2',
+    'ocdcea_e': 'C18H33O2',
+    'ocdcea_p': 'C18H33O2',
+    'ocdceap': 'C18H34O5P',
+    'octa': 'C8H15O2',
+    'octa_e': 'C8H15O2',
+    'octa_p': 'C8H15O2',
+    'octapb': 'C8H15O',
+    'octdp': 'C40H65O7P2',
+    'octeACP': 'C29H53N2O8PRS',
+    'od2coa': 'C39H64N7O17P3S',
+    'odecoa': 'C39H64N7O17P3S',
+    'ogmeACP': 'C17H27N2O11PRS',
+    'ohpb': 'C4H4O8P',
+    'op4en': 'C5H5O3',
+    'opmeACP': 'C19H31N2O11PRS',
+    'orn': 'C5H13N2O2',
+    'orn_e': 'C5H13N2O2',
+    'orn_p': 'C5H13N2O2',
+    'orot': 'C5H3N2O4',
+    'orot5p': 'C10H10N2O11P',
+    'orot_e': 'C5H3N2O4',
+    'orot_p': 'C5H3N2O4',
+    'oxa': 'C2O4',
+    'oxadpcoa': 'C27H37N7O20P3S',
+    'oxalcoa': 'C23H31N7O19P3S',
+    'oxam': 'C2H2NO3',
+    'oxur': 'C3H3N2O4',
+    'pa120': 'C27H51O8P1',
+    'pa120_p': 'C27H51O8P1',
+    'pa140': 'C31H59O8P1',
+    'pa140_p': 'C31H59O8P1',
+    'pa141': 'C31H55O8P1',
+    'pa141_p': 'C31H55O8P1',
+    'pa160': 'C35H67O8P1',
+    'pa160_p': 'C35H67O8P1',
+    'pa161': 'C35H63O8P1',
+    'pa161_p': 'C35H63O8P1',
+    'pa180': 'C39H75O8P1',
+    'pa180_p': 'C39H75O8P1',
+    'pa181': 'C39H71O8P1',
+    'pa181_p': 'C39H71O8P1',
+    'pac': 'C8H7O2',
+    'pacald': 'C8H8O',
+    'pacald_e': 'C8H8O',
+    'pacald_p': 'C8H8O',
+    'palmACP': 'C27H51N2O8PRS',
+    'pan4p': 'C11H21N2O7PS',
+    'pant_R': 'C6H11O4',
+    'pap': 'C10H11N5O10P2',
+    'paps': 'C10H11N5O13P2S',
+    'pdx5p': 'C8H10NO6P',
+    'pe120': 'C29H58N1O8P1',
+    'pe120_p': 'C29H58N1O8P1',
+    'pe140': 'C33H66N1O8P1',
+    'pe140_p': 'C33H66N1O8P1',
+    'pe141': 'C33H62N1O8P1',
+    'pe141_p': 'C33H62N1O8P1',
+    'pe160': 'C37H74N1O8P1',
+    'pe160_p': 'C37H74N1O8P1',
+    'pe161': 'C37H70N1O8P1',
+    'pe161_p': 'C37H70N1O8P1',
+    'pe180': 'C41H82N1O8P1',
+    'pe180_p': 'C41H82N1O8P1',
+    'pe181': 'C41H78N1O8P1',
+    'pe181_p': 'C41H78N1O8P1',
+    'peamn_e': 'C8H12N',
+    'peamn_p': 'C8H12N',
+    'pep': 'C3H2O6P',
+    'pg120': 'C30H58O10P1',
+    'pg120_p': 'C30H58O10P1',
+    'pg140': 'C34H66O10P1',
+    'pg140_p': 'C34H66O10P1',
+    'pg141': 'C34H62O10P1',
+    'pg141_p': 'C34H62O10P1',
+    'pg160': 'C38H74O10P1',
+    'pg160_p': 'C38H74O10P1',
+    'pg161': 'C38H70O10P1',
+    'pg161_p': 'C38H70O10P1',
+    'pg180': 'C42H82O10P1',
+    'pg180_p': 'C42H82O10P1',
+    'pg181': 'C42H78O10P1',
+    'pg181_p': 'C42H78O10P1',
+    'pgp120': 'C30H57O13P2',
+    'pgp120_p': 'C30H57O13P2',
+    'pgp140': 'C34H65O13P2',
+    'pgp140_p': 'C34H65O13P2',
+    'pgp141': 'C34H61O13P2',
+    'pgp141_p': 'C34H61O13P2',
+    'pgp160': 'C38H73O13P2',
+    'pgp160_p': 'C38H73O13P2',
+    'pgp161': 'C38H69O13P2',
+    'pgp161_p': 'C38H69O13P2',
+    'pgp180': 'C42H81O13P2',
+    'pgp180_p': 'C42H81O13P2',
+    'pgp181': 'C42H77O13P2',
+    'pgp181_p': 'C42H77O13P2',
+    'phaccoa': 'C29H38N7O17P3S',
+    'phe_L': 'C9H11NO2',
+    'phe_L_e': 'C9H11NO2',
+    'phe_L_p': 'C9H11NO2',
+    'pheme': 'C34H30FeN4O4',
+    'pheme_e': 'C34H30FeN4O4',
+    'pheme_p': 'C34H30FeN4O4',
+    'phetrna': 'C9H10NOR',
+    'phhlipa': 'C124H219N2O54P3',
+    'phom': 'C4H8NO6P',
+    'phphhlipa': 'C131H230N2O63P4',
+    'phpyr': 'C9H7O3',
+    'phthr': 'C4H8NO7P',
+    'pi': 'HO4P',
+    'pi_e': 'HO4P',
+    'pi_p': 'HO4P',
+    'pimACP': 'C18H31N2O10PRS',
+    'pmeACP': 'C19H33N2O10PRS',
+    'pmtcoa': 'C37H62N7O17P3S',
+    'pnto_R': 'C9H16NO5',
+    'pnto_R_e': 'C9H16NO5',
+    'pnto_R_p': 'C9H16NO5',
+    'poaac': 'C3H5NO3',
+    'ppa': 'C3H5O2',
+    'ppa_e': 'C3H5O2',
+    'ppa_p': 'C3H5O2',
+    'ppal': 'C3H6O',
+    'ppal_e': 'C3H6O',
+    'ppal_p': 'C3H6O',
+    'ppap': 'C3H5O5P',
+    'ppbng': 'C10H13N2O4',
+    'ppcoa': 'C24H36N7O17P3S',
+    'ppgpp': 'C10H11N5O17P4',
+    'pphn': 'C10H8O6',
+    'ppi': 'HO7P2',
+    'ppp9': 'C34H32N4O4',
+    'pppg9': 'C34H38N4O4',
+    'pppi': 'HO10P3',
+    'pppn': 'C9H9O2',
+    'pppn_e': 'C9H9O2',
+    'pppn_p': 'C9H9O2',
+    'ppt_e': 'HO3P',
+    'ppt_p': 'HO3P',
+    'pram': 'C5H11NO7P',
+    'pran': 'C12H13NO9P',
+    'prbamp': 'C15H19N5O14P2',
+    'prbatp': 'C15H19N5O20P4',
+    'preq0': 'C7H5N5O',
+    'preq1': 'C7H10N5O',
+    'prfp': 'C15H21N5O15P2',
+    'prlp': 'C15H21N5O15P2',
+    'pro_L': 'C5H9NO2',
+    'pro_L_e': 'C5H9NO2',
+    'pro_L_p': 'C5H9NO2',
+    'progly': 'C7H12N2O3',
+    'progly_e': 'C7H12N2O3',
+    'progly_p': 'C7H12N2O3',
+    'protrna': 'C5H8NOR',
+    'prpp': 'C5H8O14P3',
+    'ps120': 'C30H57N1O10P1',
+    'ps140': 'C34H65N1O10P1',
+    'ps141': 'C34H61N1O10P1',
+    'ps160': 'C38H73N1O10P1',
+    'ps161': 'C38H69N1O10P1',
+    'ps180': 'C42H81N1O10P1',
+    'ps181': 'C42H77N1O10P1',
+    'psclys': 'C12H25N2O7',
+    'psclys_e': 'C12H25N2O7',
+    'psclys_p': 'C12H25N2O7',
+    'pser_L': 'C3H6NO6P',
+    'pser_L_e': 'C3H6NO6P',
+    'pser_L_p': 'C3H6NO6P',
+    'ptrc': 'C4H14N2',
+    'ptrc_e': 'C4H14N2',
+    'ptrc_p': 'C4H14N2',
+    'pyam5p': 'C8H12N2O5P',
+    'pydam': 'C8H13N2O2',
+    'pydam_e': 'C8H13N2O2',
+    'pydam_p': 'C8H13N2O2',
+    'pydx': 'C8H9NO3',
+    'pydx5p': 'C8H8NO6P',
+    'pydx_e': 'C8H9NO3',
+    'pydx_p': 'C8H9NO3',
+    'pydxn': 'C8H11NO3',
+    'pydxn_e': 'C8H11NO3',
+    'pydxn_p': 'C8H11NO3',
+    'pyr': 'C3H3O3',
+    'pyr_e': 'C3H3O3',
+    'pyr_p': 'C3H3O3',
+    'q8': 'C49H74O4',
+    'q8h2': 'C49H76O4',
+    'quin': 'C7H12O6',
+    'quin_e': 'C7H12O6',
+    'quin_p': 'C7H12O6',
+    'quln': 'C7H3NO4',
+    'r15bp': 'C5H8O11P2',
+    'r1p': 'C5H9O8P',
+    'r5p': 'C5H9O8P',
+    'r5p_e': 'C5H9O8P',
+    'r5p_p': 'C5H9O8P',
+    'ragund': 'C69H113N1O16P2',
+    'rbflvrd': 'C17H22N4O6',
+    'rbl_L': 'C5H10O5',
+    'rdmbzi': 'C14H18N2O4',
+    'rephaccoa': 'C29H38N7O18P3S',
+    'rfamp_e': 'C43H58N4O12',
+    'rfamp_p': 'C43H58N4O12',
+    'rhcys': 'C9H17NO6S',
+    'rib_D': 'C5H10O5',
+    'rib_D_e': 'C5H10O5',
+    'rib_D_p': 'C5H10O5',
+    'ribflv': 'C17H20N4O6',
+    'rml': 'C6H12O5',
+    'rml1p': 'C6H11O8P',
+    'rmn': 'C6H12O5',
+    'rmn_e': 'C6H12O5',
+    'rmn_p': 'C6H12O5',
+    'ru5p_D': 'C5H9O8P',
+    'ru5p_L': 'C5H9O8P',
+    's17bp': 'C7H12O13P2',
+    's7p': 'C7H13O10P',
+    'sarcs': 'C3H7NO2',
+    'sbt6p': 'C6H13O9P',
+    'sbt_D_e': 'C6H14O6',
+    'sbt_D_p': 'C6H14O6',
+    'sbzcoa': 'C32H39N7O20P3S',
+    'scl': 'C42H39N4O16',
+    'sectrna': 'C3H6NOSeR',
+    'sel': 'O4Se',
+    'sel_e': 'O4Se',
+    'sel_p': 'O4Se',
+    'seln': 'HSe',
+    'selnp': 'H2O3PSe',
+    'ser_D': 'C3H7NO3',
+    'ser_D_e': 'C3H7NO3',
+    'ser_D_p': 'C3H7NO3',
+    'ser_L': 'C3H7NO3',
+    'ser_L_e': 'C3H7NO3',
+    'ser_L_p': 'C3H7NO3',
+    'seramp': 'C13H19N6O9P',
+    'sertrna': 'C3H6NO2R',
+    'sertrna_LSQBKT_sec_RSQBKT_': 'C3H6NO2R',
+    'Sfglutth': 'C11H16N3O7S',
+    'sheme': 'C42H36FeN4O16',
+    'skm': 'C7H9O5',
+    'skm5p': 'C7H8O8P',
+    'skm_e': 'C7H9O5',
+    'skm_p': 'C7H9O5',
+    'sl26da': 'C11H16N2O7',
+    'sl2a6o': 'C11H12NO8',
+    'slnt': 'O3Se',
+    'slnt_e': 'O3Se',
+    'slnt_p': 'O3Se',
+    'so2': 'O2S',
+    'so2_e': 'O2S',
+    'so2_p': 'O2S',
+    'so3': 'O3S',
+    'so3_e': 'O3S',
+    'so3_p': 'O3S',
+    'so4': 'O4S',
+    'so4_e': 'O4S',
+    'so4_p': 'O4S',
+    'spmd': 'C7H22N3',
+    'spmd_e': 'C7H22N3',
+    'spmd_p': 'C7H22N3',
+    'stcoa': 'C39H66N7O17P3S',
+    'suc6p': 'C12H21O14P',
+    'sucarg': 'C10H17N4O5',
+    'sucbz': 'C11H8O5',
+    'succ': 'C4H4O4',
+    'succ_e': 'C4H4O4',
+    'succ_p': 'C4H4O4',
+    'succoa': 'C25H35N7O19P3S',
+    'sucglu': 'C9H10NO7',
+    'sucgsa': 'C9H11NO6',
+    'suchms': 'C8H12NO6',
+    'sucorn': 'C9H15N2O5',
+    'sucr_e': 'C12H22O11',
+    'sucr_p': 'C12H22O11',
+    'sucsal': 'C4H5O3',
+    'sufbcd': 'H8O2S6R',
+    'sufbcd_2fe2s': 'H4O2S8Fe2R',
+    'sufbcd_2fe2s2': 'O2S10Fe4R',
+    'sufbcd_4fe4s': 'H4O2S10Fe4R',
+    'sufse': 'HSR',
+    'sufsesh': 'HS2R',
+    'sulfac': 'C2H2O5S',
+    'sulfac_e': 'C2H2O5S',
+    'sulfac_p': 'C2H2O5S',
+    't3c11vaceACP': 'C29H51N2O8PRS',
+    't3c5ddeceACP': 'C23H39N2O8PRS',
+    't3c7mrseACP': 'C25H43N2O8PRS',
+    't3c9palmeACP': 'C27H47N2O8PRS',
+    'tag6p_D': 'C6H11O9P',
+    'tagdp_D': 'C6H10O12P2',
+    'tagur': 'C6H9O7',
+    'tartr_D': 'C4H4O6',
+    'tartr_D_e': 'C4H4O6',
+    'tartr_D_p': 'C4H4O6',
+    'tartr_L': 'C4H4O6',
+    'tartr_L_e': 'C4H4O6',
+    'tartr_L_p': 'C4H4O6',
+    'taur': 'C2H7NO3S',
+    'taur_e': 'C2H7NO3S',
+    'taur_p': 'C2H7NO3S',
+    'tcynt': 'CNS',
+    'tcynt_e': 'CNS',
+    'tcynt_p': 'CNS',
+    'td2coa': 'C35H56N7O17P3S',
+    'tdcoa': 'C35H58N7O17P3S',
+    'tddec2eACP': 'C23H41N2O8PRS',
+    'tdeACP': 'C25H45N2O8PRS',
+    'tdec2eACP': 'C21H37N2O8PRS',
+    'tdecoa': 'C35H56N7O17P3S',
+    'thdp': 'C7H7NO4',
+    'thex2eACP': 'C17H29N2O8PRS',
+    'thf': 'C19H21N7O6',
+    'thm': 'C12H17N4OS',
+    'thm_e': 'C12H17N4OS',
+    'thm_p': 'C12H17N4OS',
+    'thmmp': 'C12H16N4O4PS',
+    'thmnp': 'C9H15N5O4',
+    'thmpp': 'C12H16N4O7P2S',
+    'thptdn': 'C6H10N4',
+    'thr_L': 'C4H9NO3',
+    'thr_L_e': 'C4H9NO3',
+    'thr_L_p': 'C4H9NO3',
+    'thrp_e': 'C4H8NO6P',
+    'thrp_p': 'C4H8NO6P',
+    'thrtrna': 'C4H8NO2R',
+    'thym': 'C5H6N2O2',
+    'thym_e': 'C5H6N2O2',
+    'thym_p': 'C5H6N2O2',
+    'thymd': 'C10H14N2O5',
+    'thymd_e': 'C10H14N2O5',
+    'thymd_p': 'C10H14N2O5',
+    'tma': 'C3H10N',
+    'tma_e': 'C3H10N',
+    'tma_p': 'C3H10N',
+    'tmao': 'C3H9NO',
+    'tmao_e': 'C3H9NO',
+    'tmao_p': 'C3H9NO',
+    'tmrs2eACP': 'C25H45N2O8PRS',
+    'toct2eACP': 'C19H33N2O8PRS',
+    'toctd2eACP': 'C29H53N2O8PRS',
+    'tpalm2eACP': 'C27H49N2O8PRS',
+    'trdox': 'X',
+    'trdrd': 'XH2',
+    'tre': 'C12H22O11',
+    'tre6p': 'C12H21O14P',
+    'tre_e': 'C12H22O11',
+    'tre_p': 'C12H22O11',
+    'trnaala': 'R',
+    'trnaarg': 'R',
+    'trnaasn': 'C10H17O10PR2',
+    'trnaasp': 'R',
+    'trnacys': 'R',
+    'trnagln': 'R',
+    'trnaglu': 'R',
+    'trnagly': 'R',
+    'trnahis': 'R',
+    'trnaile': 'R',
+    'trnaleu': 'R',
+    'trnalys': 'R',
+    'trnamet': 'R',
+    'trnaphe': 'R',
+    'trnapro': 'R',
+    'trnasecys': 'R',
+    'trnaser': 'R',
+    'trnathr': 'R',
+    'trnatrp': 'R',
+    'trnatyr': 'R',
+    'trnaval': 'R',
+    'trp_L': 'C11H12N2O2',
+    'trp_L_e': 'C11H12N2O2',
+    'trp_L_p': 'C11H12N2O2',
+    'trptrna': 'C11H11N2OR',
+    'tsul': 'O3S2',
+    'tsul_e': 'O3S2',
+    'tsul_p': 'O3S2',
+    'ttdca': 'C14H27O2',
+    'ttdca_e': 'C14H27O2',
+    'ttdca_p': 'C14H27O2',
+    'ttdcap': 'C14H28O5P',
+    'ttdcea': 'C14H25O2',
+    'ttdcea_e': 'C14H25O2',
+    'ttdcea_p': 'C14H25O2',
+    'ttdceap': 'C14H26O5P',
+    'ttrcyc_e': 'C22H24N2O8',
+    'ttrcyc_p': 'C22H24N2O8',
+    'tungs': 'O4W1',
+    'tungs_e': 'O4W1',
+    'tungs_p': 'O4W1',
+    'tym_e': 'C8H12NO',
+    'tym_p': 'C8H12NO',
+    'tyr_L': 'C9H11NO3',
+    'tyr_L_e': 'C9H11NO3',
+    'tyr_L_p': 'C9H11NO3',
+    'tyrp_e': 'C9H10NO6P',
+    'tyrp_p': 'C9H10NO6P',
+    'tyrtrna': 'C9H10NO2R',
+    'u23ga': 'C43H75N3O20P2',
+    'u3aga': 'C31H51N3O19P2',
+    'u3hga': 'C29H50N3O18P2',
+    'uaagmda': 'C95H152N8O28P2',
+    'uaccg': 'C20H26N3O19P2',
+    'uacgam': 'C17H25N3O17P2',
+    'uacgam_e': 'C17H25N3O17P2',
+    'uacgam_p': 'C17H25N3O17P2',
+    'uacmam': 'C17H25N3O17P2',
+    'uacmamu': 'C17H22N3O18P2',
+    'uagmda': 'C87H139N7O23P2',
+    'uama': 'C23H33N4O20P2',
+    'uamag': 'C28H39N5O23P2',
+    'uamr': 'C20H28N3O19P2',
+    'udcpdp': 'C55H89O7P2',
+    'udcpdp_p': 'C55H89O7P2',
+    'udcpp': 'C55H89O4P',
+    'udcpp_p': 'C55H89O4P',
+    'udp': 'C9H11N2O12P2',
+    'udpacgal_e': 'C17H25N3O17P2',
+    'udpacgal_p': 'C17H25N3O17P2',
+    'udpg': 'C15H22N2O17P2',
+    'udpg_e': 'C15H22N2O17P2',
+    'udpg_p': 'C15H22N2O17P2',
+    'udpgal': 'C15H22N2O17P2',
+    'udpgal_e': 'C15H22N2O17P2',
+    'udpgal_p': 'C15H22N2O17P2',
+    'udpgalfur': 'C15H22N2O17P2',
+    'udpglcur': 'C15H19N2O18P2',
+    'udpglcur_e': 'C15H19N2O18P2',
+    'udpglcur_p': 'C15H19N2O18P2',
+    'udpLa4fn': 'C15H21N3O16P2',
+    'udpLa4n': 'C14H22N3O15P2',
+    'udpLa4o': 'C14H18N2O16P2',
+    'ugmd': 'C35H51N7O26P2',
+    'ugmda': 'C41H61N9O28P2',
+    'uLa4fn': 'C61H99N1O8P1',
+    'uLa4n': 'C60H100N1O7P1',
+    'uLa4n_p': 'C60H100N1O7P1',
+    'um4p': 'C38H56N8O27P2',
+    'ump': 'C9H11N2O9P',
+    'ump_e': 'C9H11N2O9P',
+    'ump_p': 'C9H11N2O9P',
+    'unaga': 'C63H103NO12P2',
+    'unagamu': 'C71H113N2O18P2',
+    'unagamuf': 'C79H126N3O22P2',
+    'unagamuf_p': 'C79H126N3O22P2',
+    'uppg3': 'C40H36N4O16',
+    'ura': 'C4H4N2O2',
+    'ura_e': 'C4H4N2O2',
+    'ura_p': 'C4H4N2O2',
+    'uracp': 'C4H6N2O4',
+    'urate': 'C5H4N4O3',
+    'urdglyc': 'C3H5N2O4',
+    'urea': 'CH4N2O',
+    'urea_e': 'CH4N2O',
+    'urea_p': 'CH4N2O',
+    'uri': 'C9H12N2O6',
+    'uri_e': 'C9H12N2O6',
+    'uri_p': 'C9H12N2O6',
+    'utp': 'C9H11N2O15P3',
+    'val_L': 'C5H11NO2',
+    'val_L_e': 'C5H11NO2',
+    'val_L_p': 'C5H11NO2',
+    'valtrna': 'C5H10NOR',
+    'wco': 'C10H10N5O8PS2W',
+    'xan': 'C5H4N4O2',
+    'xan_e': 'C5H4N4O2',
+    'xan_p': 'C5H4N4O2',
+    'xdp': 'C10H11N4O12P2',
+    'xmp': 'C10H11N4O9P',
+    'xmp_e': 'C10H11N4O9P',
+    'xmp_p': 'C10H11N4O9P',
+    'xtp': 'C10H11N4O15P3',
+    'xtsn': 'C10H12N4O6',
+    'xtsn_e': 'C10H12N4O6',
+    'xtsn_p': 'C10H12N4O6',
+    'xu5p_D': 'C5H9O8P',
+    'xu5p_L': 'C5H9O8P',
+    'xyl_D': 'C5H10O5',
+    'xyl_D_e': 'C5H10O5',
+    'xyl_D_p': 'C5H10O5',
+    'xylu_D': 'C5H10O5',
+    'xylu_L': 'C5H10O5',
+    'xylu_L_e': 'C5H10O5',
+    'xylu_L_p': 'C5H10O5',
+    'zn2': 'Zn',
+    'zn2_e': 'Zn',
+    'zn2_p': 'Zn'
+}
