+++ code/core/sbmlio.py
@@ -6,10 +6,13 @@ The sbmlio module provides support for input/output of ReactionNetwork structure
 """
 
 import re
-import core, DB, enhancedLists, labeling
+import DB, enhancedLists, labeling
+import core
 import Genes, Proteins
-import utilities as utils
+from SBMLResourceIdentifiers import SBMLResourceIdentifiers
+import utilities
 import libsbml
+
 
 
 class SBMLImporter:
@@ -21,8 +24,10 @@ class SBMLImporter:
         self.model_name = None
         self.model_notes = {}
 
-        self.metList = None
+        self.metList      = None
         self.reactionList = None
+        self.proteinSet   = None
+        self.geneSet      = None
 
         self.initialSetFluxesCount = None
         self.initialBoundedFluxesCount = None
@@ -47,14 +52,45 @@ class SBMLImporter:
                 for x in range(self.errorCount):
                     err = errs.getError(x)
                     self.logStrings.append(err.getMessage())
-            self.parseSBMLDoc(sbmlDoc)
 
         if self.errorCount > 0 and exceptionOnError:
             raise Exception('Errors parsing SBML:\n%s' % ('\n'.join(self.logStrings)))
+        else:
+            self.parseSBMLDoc(sbmlDoc)
+
 
 
     def parseSBMLDoc(self, sbmlDoc):
+
+        hasFBC = False
+        usingFBC = False
         model = sbmlDoc.getModel()
+
+        # Basic check for FBC functionality
+        if hasattr(libsbml, 'SBML_FBC_GENEPRODUCT'):
+            if libsbml.SBML_FBC_GENEPRODUCT:
+                hasFBC = True
+
+        if not hasFBC:
+            self.logStrings.append('Warning: libsbml does not appear to have FBC (flux balance constraints) extension compiled in.')
+
+        modelFBCplugin = None
+        FBCversion = 0
+        FBCstrict = True
+
+        if hasFBC:
+            modelFBCplugin = model.getPlugin(str("fbc"))
+            if modelFBCplugin is not None:
+                usingFBC = True
+                FBCversion = modelFBCplugin.getPackageVersion()
+                if FBCversion == 2:
+                    FBCstrict = modelFBCplugin.getStrict()
+                    if not FBCstrict:
+                        self.logStrings.append('Warning: This model is using the FBC package in non-strict mode.  Strange things may occur; see the libsbml docs.')
+
+        masterProteinSet = Proteins.ProteinSet()
+        masterGeneSet = Genes.GeneSet()
+
         self.model_name = model.getName()
         self.logStrings.append('Parsed model name ' + self.model_name)
 
@@ -79,8 +115,10 @@ class SBMLImporter:
         mets = []
         anyFeed   = False
         anySource = False
+        metIdentifiersCount = 0
 
         for sp in listOfSpecies:
+
             # Changing met name to appropiate form
             species_id  = re.sub('^M_','', sp.getId()) # ID must exist in valid SBML   
             full_name   = sp.getName()
@@ -98,6 +136,13 @@ class SBMLImporter:
             destination = False
             carbonsOut  = None
             formula     = ''
+
+            spFbc = None
+            if usingFBC:
+                spFbc = sp.getPlugin("fbc")
+
+            if usingFBC and spFbc:
+                formula = spFbc.getChemicalFormula()
 
             # Note: Converts all names to SBML
             spName     = core.MetaboliteName(species_id)
@@ -118,14 +163,65 @@ class SBMLImporter:
 
                 anySource = True if source else anySource
                 anyFeed   = True if feed else anyFeed
-          
-            met = core.Metabolite(spName, ncarbons, excluded, source, feed, destination, carbonsOut, full_name=full_name, formula=formula, extraNotes=spNotes)
+
+            midentifiers = SBMLResourceIdentifiers()
+            if usingFBC and (FBCversion == 2):
+                terms = self.readCVTerms(sp)
+                if 'is' in terms:
+                    metIdentifiersCount += len(terms['is'])
+                    midentifiers.add(terms['is'])
+                    #self.logStrings.append('  Met identifiers: ' + str(midentifiers))
+
+            met = core.Metabolite(spName, ncarbons, excluded, source, feed, destination, carbonsOut,
+                                  full_name=full_name, formula=formula,
+                                  extraNotes=spNotes, identifiers=midentifiers)
             mets.append(met)
 
-        self.logStrings.append('Detected ' + str(len(listOfSpecies)) + ' species, parsed ' + str(len(mets)) + ' metabolites')
-            
+        self.logStrings.append('Detected %s species, parsed %s metabolites with %s identifiers.' %
+                                (str(len(listOfSpecies)), str(len(mets)), str(metIdentifiersCount)))
+
         # Creating met list    
         metList = enhancedLists.MetaboliteList(mets)
+
+        if (modelFBCplugin is not None) and (FBCversion == 2):
+            for gpIndex in range(modelFBCplugin.getNumGeneProducts()):
+                oneProduct = modelFBCplugin.getGeneProduct(gpIndex)
+
+                # fbc:geneProduct elements have nultiple sources for names:
+                #
+                #  fbc:label (example: 'b0584')
+                #  fbc:name  (example: 'fepA')
+                #  metaid    (example: 'G_b0584')
+                #  fbc:id    (example: 'G_b0584')
+                #
+                # The canonical name is fbc:name, and the name used to resolve against
+                # other structures can be metaid or fbc:id, depending on the structure.
+                # For our purposes, fbc:id is the relevant one.
+                #
+                # Some geneProducts do not have an fbc:name, in which case we just use fbc:id.
+
+                pid = oneProduct.getId()
+                pid = pid.replace('G_','',1)
+                # Since we specify the fbc:name first, it becomes the canonicalName:
+                n = oneProduct.getName()
+                if n:
+                    newProtein = Proteins.Protein([n, pid])
+                else:
+                    newProtein = Proteins.Protein([pid])
+                    newProtein.nonCanonicalName = True
+
+                newProtein.label = oneProduct.getLabel()
+                newProtein.sbo = oneProduct.getSBOTermID()
+                newGene = Genes.Gene(pid)
+                newProtein.encodedBy = [newGene]
+
+                terms = self.readCVTerms(oneProduct)
+                if 'is' in terms:
+                    newProtein.addIdentifier(terms['is'])
+                if 'isEncodedBy' in terms:
+                    newGene.addIdentifier(terms['isEncodedBy'])
+                masterProteinSet.recastItems(newProtein)
+                masterGeneSet.recastItems(newGene)
 
         # Warning if no feed
         if not anyFeed and anySource:
@@ -141,10 +237,21 @@ class SBMLImporter:
         closely_bounded_fluxes_count = 0
 
         reactions = []
+        rxnIdentifiersCount = 0
         for rxn in listOfReactions:
-            rxnNameOrig     = rxn.getId().replace('R_','',1).replace('_e_','(e)',1)
+            rxnNameOrig = re.sub('^R_','', rxn.getId())
+            rxnNameOrig = re.sub('_e_$','(e)', rxnNameOrig)
             rxnName         = core.ReactionName(rxnNameOrig)
+            full_name   = rxn.getName()
+            if not isinstance(full_name, str):
+                full_name = rxnName
+            elif full_name.strip() == '':
+                full_name = rxnName
+
             exchangeReaction = True if rxnName[0:3] == 'EX_' else False 
+
+            if usingFBC:
+                reactionFBC = rxn.getPlugin('fbc')
 
             # Stoichiometry
             sbml_reactants = rxn.getListOfReactants()
@@ -167,11 +274,13 @@ class SBMLImporter:
             #  Carbon Transitions and other notes
             transitionLine = 'None'
             subsystem = 'None'
+
             flux = None
             geneString = None
             geneValueString = None
             proteinString = None
             proteinValueString = None
+
             rxnNotes = {}
             if rxn.isSetNotes():
                 rxnNotesSBML = rxn.getNotes()
@@ -180,9 +289,10 @@ class SBMLImporter:
                     tl = rxnNotes['CARBON_TRANSITIONS'].replace('==','<==>').replace('--','-->')
                     transitionsLine = core.AtomTransition(rxnName + "\t" + tl)
                     transitionLine = transitionsLine.fullLine
+                if 'FLUX_VALUE' in rxnNotes:
+                    flux = rxnNotes['FLUX_VALUE']
                 if 'FLUX VALUE' in rxnNotes:
                     flux = rxnNotes['FLUX VALUE']
-                    set_fluxes_count += 1            
                 if 'SUBSYSTEM' in rxnNotes:
                     subsystem = rxnNotes['SUBSYSTEM']
                 if 'GENE_ASSOCIATION' in rxnNotes:
@@ -195,6 +305,31 @@ class SBMLImporter:
                     proteinValueString = rxnNotes['PROTEIN_COPY_VALUES']
 
             geneSets = []
+            proteinSets = []
+            # We will recast the sets we make into identical-looking sets that are composed of
+            # already-existing gene/protein objects, if available.
+            recastGeneSets = []
+            recastProteinSets = []
+
+            # If using FBC package, look for gene product associations in FBC structures.
+            if usingFBC:
+                gpas = reactionFBC.getGeneProductAssociation()
+                if gpas != None:
+                    gpasa = gpas.getAssociation()
+                    if gpasa != None:
+                        self.readGeneProductRefs(gpasa, proteinSets)
+                        for subSet in proteinSets:
+                            recastProteinSets.append(masterProteinSet.recastSet(subSet, mergeValues=True))
+                        # Use the 'encodedBy' information read from the global set
+                        # (if it exists) to build a list of correlated genes
+                        for ps in recastProteinSets:
+                            gs = []
+                            for p in ps.itemList():
+                                for gn in p.encodedBy:
+                                    gs.append(gn)
+                            if len(gs) > 0:
+                                geneSets.append(Genes.GeneSet(gs))
+
             if geneString:
                 geneString = re.sub('^\s*:','', geneString) # Chop off any bogus colon
                 geneString = geneString.strip()
@@ -202,20 +337,54 @@ class SBMLImporter:
                     geneString = ""
                 if geneString != "":
                     geneSets = Genes.GeneSet.createSetsFromStrings(geneString, geneValueString)
-            proteinSets = []
             if proteinString:
                 proteinString = re.sub('^\s*:','', proteinString) # Chop off any bogus colon
                 proteinString = proteinString.strip()
                 if proteinString.lower() in ["none", "n.a.", "n/a"]:
                     proteinString = ""
                 if proteinString != "":
+                    # The gene names act as alternate names fro the proteins here.
+                    # (We assume they are both specified in the same order, and link them up.)
+                    # This is done because at least one SBML exporter (it is not known which) erroneously pair the
+                    # transcription value with the protein identifier.
                     proteinSets = Proteins.ProteinSet.createSetsFromStrings(proteinString, geneString, proteinValueString)
+                    for subSet in proteinSets:
+                        recastProteinSets.append(masterProteinSet.recastSet(subSet, mergeValues=True))
+
+            for subSet in geneSets:
+                recastGeneSets.append(masterGeneSet.recastSet(subSet, mergeValues=True))
 
             #  Measured fluxes (MeasFluxes for 13CMFA, UB and LB for 2S-13CMFA or FBA)
             cval = 0
             LB = None
             UB = None
 
+            # If using FBC package, attempt to locate and then do some rudimentary translation
+            # on the upper and lower bounds.
+            if usingFBC:
+                LB = reactionFBC.getLowerFluxBound()
+                UB = reactionFBC.getUpperFluxBound()
+                bounds_translation = {
+                    'cobra_0_bound': '0',
+                    'cobra_default_lb': core.DEFAULT_LOWER_BOUND,
+                    'cobra_default_ub': core.DEFAULT_UPPER_BOUND
+                }
+                if LB is not None:
+                    if LB in bounds_translation:
+                        LB = bounds_translation[LB]
+                    # TODO: Right now if it's not a number we don't know what to do with it, so we set it to 0.
+                    # This is bad.  It could be something like 'R_ATPM_lower_bound', 'R_EX_cbl1_e_lower_bound', or
+                    # anything else defined in the <listOfParameters> section of the SBML document...
+                    elif not utilities.is_float(LB):
+                        LB = core.DEFAULT_LOWER_BOUND
+
+                if UB is not None:
+                    if UB in bounds_translation:
+                        UB = bounds_translation[UB]
+                    elif not utilities.is_float(UB):
+                        LB = core.DEFAULT_UPPER_BOUND
+
+            # If we have a kinetic law annotated with old-style flux bounds, we'll prefer that over the FBC package.
             kLaw = rxn.getKineticLaw()
             if kLaw:
                 listOfParameters = kLaw.getListOfParameters()
@@ -224,35 +393,120 @@ class SBMLImporter:
                         LB = par.getValue()
                     elif par.getId() == 'UPPER_BOUND':
                         UB = par.getValue()
+                    elif par.getId() == 'FLUX_VALUE':
+                        if flux is None:    # Prefer the one from the notes
+                            flux = core.flux(for_back_tup=('NA','NA'),net_exc_tup=(par.getValue(),'NA'))
+                            # Sigh...  Yet another place where we're going to have to remove 'NA' and replace it with something
+                            # sane...
                     elif par.getId() == 'OBJECTIVE_COEFFICIENT':
                         cval = par.getValue()
                         Cvec[rxnName] = par.getValue()
-                if (LB != None) and (UB != None):
-                    bounded_fluxes_count += 1
-                    if (LB == UB):
-                        closely_bounded_fluxes_count += 1
-                
+
+            if (LB != None) and (UB != None):
+                bounded_fluxes_count += 1
+                if (LB == UB):
+                    closely_bounded_fluxes_count += 1
+
             measured,fluxBounds = core.fluxBounds(LB, UB, rxn.getReversible(), exchangeReaction)
+
+            if flux is not None:
+                set_fluxes_count += 1
 
             # If we don't have a determined flux, use the flux bounds.
             if not flux:
                 flux = fluxBounds
 
+            ridentifiers = SBMLResourceIdentifiers()
+            if usingFBC and (FBCversion == 2):
+                terms = self.readCVTerms(rxn)
+                if 'is' in terms:
+                    rxnIdentifiersCount += len(terms['is'])
+                    ridentifiers.add(terms['is'])
+                    #self.logStrings.append('  React identifiers: ' + str(ridentifiers))
+
             # Instantiating reaction
             new_reaction = core.Reaction(rxnName, rxn.getReversible(), False, measured, flux=flux, fluxBounds=fluxBounds,
                                      transitionLine=transitionLine, exchange=exchangeReaction, reactants=reacts, products=prods,
-                                     cval=cval, subsystem=subsystem, gene=geneSets, protein=proteinSets, extraNotes=rxnNotes)
+                                     cval=cval, subsystem=subsystem, gene=recastGeneSets, protein=recastProteinSets,
+                                     extraNotes=rxnNotes, identifiers=ridentifiers, full_name=full_name)
             reactions.append(new_reaction)
 
         self.metList      = metList
         self.reactionList = enhancedLists.ReactionList(reactions)
+        self.proteinSet   = masterProteinSet
+        self.geneSet      = masterGeneSet
 
         self.initialSetFluxesCount     = set_fluxes_count
         self.initialBoundedFluxesCount = bounded_fluxes_count
         self.initialCloselyBoundedFluxesCount = closely_bounded_fluxes_count
 
-        self.logStrings.append('Detected ' + str(len(listOfReactions)) + ' reactions, parsed ' + str(len(reactions)) + '.')
-        self.logStrings.append('Found ' + str(bounded_fluxes_count) + ' bounded fluxes (with ' + str(closely_bounded_fluxes_count) + ' closely bounded) and ' + str(set_fluxes_count) + ' set fluxes.')
+        self.logStrings.append('Detected %s reactions, parsed %s, with %s identifiers.' %
+                                (str(len(listOfReactions)), str(len(reactions)), str(rxnIdentifiersCount)))
+        self.logStrings.append('Found %s bounded fluxes (with %s closely bounded) and %s set fluxes.' %
+                                (str(bounded_fluxes_count), str(closely_bounded_fluxes_count), str(set_fluxes_count)))
+        self.logStrings.append('Detected %s genes, %s proteins.' %
+                                (str(len(masterGeneSet)), str(len(masterProteinSet))))
+
+    qualifiersToElements = {
+        libsbml.BQB_ENCODES         : "encodes",        # 8
+        libsbml.BQB_HAS_PART        : "hasPart",        # 1
+        libsbml.BQB_HAS_PROPERTY    : "hasProperty",    # 10
+        libsbml.BQB_HAS_VERSION     : "hasVersion",     # 4
+        libsbml.BQB_IS              : "is",             # 0
+        libsbml.BQB_IS_DESCRIBED_BY : "isDescribedBy",  # 6
+        libsbml.BQB_IS_ENCODED_BY   : "isEncodedBy",    # 7
+        libsbml.BQB_IS_HOMOLOG_TO   : "isHomologTo",    # 5
+        libsbml.BQB_IS_PART_OF      : "isPartOf",       # 2
+        libsbml.BQB_IS_PROPERTY_OF  : "isPropertyOf",   # 11
+        libsbml.BQB_IS_VERSION_OF   : "isVersionOf",    # 3
+        libsbml.BQB_OCCURS_IN       : "occursIn",       # 9
+        libsbml.BQB_UNKNOWN         : "unknown",        # 12
+
+#        libsbml.BQM_IS              : "isB",             # 0   # Should be in another structure eventually
+#        libsbml.BQM_IS_DERIVED_FROM : "isDerivedFrom",   # 2
+#        libsbml.BQM_IS_DESCRIBED_BY : "isDescribedByB",  # 1
+#        libsbml.BQM_UNKNOWN         : "unknownB"         # 3
+    }
+
+
+    def readCVTerms(self, sbaseObj):
+        terms = {}
+        if sbaseObj.getNumCVTerms() < 1:
+            return terms
+        for i in range(sbaseObj.getNumCVTerms()):
+            oneTerm = sbaseObj.getCVTerm(i)
+            if oneTerm.getQualifierType() == libsbml.BIOLOGICAL_QUALIFIER:
+                name = SBMLImporter.qualifiersToElements[oneTerm.getBiologicalQualifierType()]
+            else:
+                name = SBMLImporter.qualifiersToElements[oneTerm.getModelQualifierType()]
+            resourceList = []
+            for j in range(oneTerm.getNumResources()):
+                uri = oneTerm.getResourceURI(j)
+                if name != None and uri != None:
+                    resourceList.append(uri)
+            terms[name] = tuple(resourceList)
+        return terms
+
+
+    # This function calls itself recursively to walk through sets of sets when
+    # indicated by the ibsbml.FbcOr element.
+    def readGeneProductRefs(self, gpasa, listContainer):
+        if isinstance(gpasa, libsbml.GeneProductRef):
+            p = gpasa.getGeneProduct()
+            listContainer.append(Proteins.ProteinSet(Proteins.Protein(p.replace('G_','',1))))
+        elif isinstance(gpasa, libsbml.FbcAnd):
+            subList = []
+            for i in range(gpasa.getNumAssociations()):
+                gpasaSub = gpasa.getAssociation(i)
+                p = gpasaSub.getGeneProduct()
+                subList.append(Proteins.Protein(p.replace('G_','',1)))
+            listContainer.append(Proteins.ProteinSet(subList))
+        elif isinstance(gpasa, libsbml.FbcOr):
+            for i in range(gpasa.getNumAssociations()):
+                gpasaSub = gpasa.getAssociation(i)
+                self.readGeneProductRefs(gpasaSub, listContainer)
+        else:
+            self.logStrings.append('Strange element structure found while parsing fbc geneProduct associations!')
 
 
 
@@ -557,7 +811,7 @@ def convertSBMLNotes(notesSBML='unset', notesDict='unset'):
             if core.rangedNumber.looks_like_rangedNumber(item):
                 item = core.rangedNumber.from_string(item)
             else:
-                if utils.is_float(item):
+                if utilities.is_float(item):
                     item = float(item)
             processed.append(item)
 
@@ -713,26 +967,32 @@ def convertSBMLNotes(notesSBML='unset', notesDict='unset'):
     raise Exception('convertSBMLNotes called with no valid input')
 
 
-##### UNIT TESTS ########
 
-import unittest
-
-# Expected result from tests below
-expectedResult =  '153.68'  
-
-
-
-class testSBML(unittest.TestCase):  
+def test():  
     """ Testing basic usage of SBML support """
+    sbmlFileName = "SBMLTestFile.sbml"
+    sbmlFileContents = None
+    with open(sbmlFileName, 'r') as file:
+        sbmlFileContents = file.read()
+    file.closed
+    assert sbmlFileContents is not None, "Could not load SBML test file %s." % sbmlFileName
 
-    def setUp(self):
-        pass
+    sbmlImporter = SBMLImporter(sbmlFileContents)
+
+    name         = sbmlImporter.model_name
+    notes        = sbmlImporter.model_notes
+    metList      = sbmlImporter.metList
+    reactionList = sbmlImporter.reactionList
+
+    errorCount   = sbmlImporter.errorCount
+    logStrings   = sbmlImporter.logStrings
+
+    if errorCount > 0 and exceptionOnError:
+        raise Exception('Errors interpreting SBML:\n%s' % ('\n'.join(logStrings)))
+    else:
+        print '\n'.join(logStrings)
 
 
-    def testSBML(self):
-        """ Testing SBML support """
-
-
-    def tearDown(self):
-        pass
+if __name__ == "__main__":
+    test()
